% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_cv.R
\name{mt_anno_cv}
\alias{mt_anno_cv}
\title{Calculate CV - coefficient of variation}
\usage{
mt_anno_cv(D, qc_samples, out_col, replicates = F, id_col = NULL)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{qc_samples}{Logical expression. Can use fields from colData.}

\item{out_col}{Name of new rowData column to store the cv values.}

\item{replicates}{OPTIONAL. Flag to indicate if the selected samples are replicates. Default: F.}

\item{id_col}{OPTIONAL. If replicates is T, it is the name of the column in colData containing sample IDs.}
}
\value{
rowData: new column added
}
\description{
{ADD DESCRIPTION}
}
\examples{
\dontrun{... \%>\% mt_anno_cv(qc_samples=="PQC", out_col = "PQC_cv") \%>\% ...}

}
\author{
AS, RB
}
