% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_icc.R
\name{mt_anno_icc}
\alias{mt_anno_icc}
\title{Calculate intraclass correlation (ICC)}
\usage{
mt_anno_icc(D, qc_samples, out_col, id_col, icc_lmer = F)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{qc_samples}{Logical expression. Can use fields from colData.}

\item{out_col}{Name of new rowData column to store the icc values.}

\item{id_col}{Name of the colData column containing sample IDs.}

\item{icc_lmer}{Optional flag to set lmer flag in ICC function default is FALSE}
}
\value{
rowData: New annotation column added.
}
\description{
{ADD DESCRIPTION}
}
\examples{
\dontrun{... \%>\% mt_anno_icc(qc_samples= (RID\%in\%RID[duplicated(RID)]),
out_col = "Dup_icc", id_col='RID') \%>\% ...}

}
\author{
RB
}
