% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_missingness.R
\name{mt_anno_missingness}
\alias{mt_anno_missingness}
\title{Missingness Annotation for Metabolites / Samples}
\usage{
mt_anno_missingness(D, anno_type, out_col)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{anno_type}{"samples" (colData) or "metabolites" (rowNames) to add missingness column to}

\item{out_col}{name of new missingness column (if missing, use "missingness")}
}
\value{
rowData or colData: new annotation column added
}
\description{
Adds a rowData or colData column representing the missingness of metabolites or samples, respectively.
}
\examples{
# example of how to run function
\dontrun{
#load data
mt_files_load_metabolon(file=file, sheet="data") \%>\%
# add metabolite missingness column to rowData
mt_anno_missingness(anno_type = "metabolites", out_col = "Metab_Missing")
# add sample missingness column to colData
mt_anno_missingness(anno_type = "samples", out_col = "Samp_Missing")
}

}
\author{
KC
}
