% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_mutate.R
\name{mt_anno_mutate}
\alias{mt_anno_mutate}
\title{Create new variable by expression}
\usage{
mt_anno_mutate(D, anno_type, col_name, term)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{anno_type}{direction, "samples" or "metabolites"}

\item{col_name}{Name of new annotation column to store result of expression.}

\item{term}{Mutate term to forward to \code{dplyr::mutate()}.}
}
\value{
colData or rowData: New annotation column added.
}
\description{
Creates a new variable using dplyr's mutate() mechanism. Formula can use any column in the respective colData/rowData.
}
\examples{
\dontrun{# Convert numeric sample annotation field 'num' to factor
...  \%>\%
 mt_anno_mutate(anno_type="samples", col_name="num", term = as.factor(num)) \%>\% ...}

}
\author{
JK
}
