% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_pathways_from_file.R
\name{mt_anno_pathways_from_file}
\alias{mt_anno_pathways_from_file}
\title{Add pathway information.}
\usage{
mt_anno_pathways_from_file(
  D,
  in_col,
  out_col,
  file,
  sheet,
  met_ID,
  pw_ID,
  pw_name,
  export_raw_db
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{in_col}{rowData column to use for pathway fetching. The selected column must contain metabolite identifiers (e.g. HMBD, KEGG, ChEBI, etc)}

\item{out_col}{new column name for rowData to output pathway information to}

\item{file}{path where the pathway annotation flat file is stored}

\item{sheet}{sheet name or number to read in flat file}

\item{met_ID}{flat file colname: this column should contain metabolite IDs}

\item{pw_ID}{flat file colname: this column should contain pathway IDs}

\item{pw_name}{flat file colname: this column should contain pathway names}

\item{export_raw_db}{OPTIONAL. Export the pathway database to a directory. Must be a string containing the path name with a .xlsx extension.}
}
\value{
rowData: new pathway annotation column added

$result$pathways: a dataframe of pathway information
}
\description{
Adds custom pathways to the already existing SummarizedExperiment
 data structure using a flat file.
 NOTE: The flat file should be an Excel file, containing 3 columns:
 met_ID, containing metabolite IDs, pw_ID containing pathway IDs,
 and pw_name containing pathway names.
 The pathway columns pw_ID and pw_name should be in long format, meaning
 if a metabolite is assigned to multiple pathways, each metabolite pathway
 pair should appear in a separate row.
}
\examples{
\dontrun{# annotate metabolites using the SMP column of the pathway database flat file
... \%>\%
      mt_anno_pathways_from_file(in_col = "HMDb_ID",
                                 out_col = "janpw",
                                 file = codes.makepath("packages/metabotools_external/hmdb/hmdb_preprocessed_4.0.csv"),
                                 met_ID = "HMDB_id",
                                 pw_ID = "SMP",
                                 pw_name = "pathway_name") \%>\%
...}

}
\author{
Parviz Gomari
}
