% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_pathways_graphite.R
\name{mt_anno_pathways_graphite}
\alias{mt_anno_pathways_graphite}
\title{Add pathway information.}
\usage{
mt_anno_pathways_graphite(
  D,
  in_col,
  out_col,
  pwdb_species,
  pwdb_name,
  n_cores = 2,
  export_raw_db
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{in_col}{rowData olumn to use for pathway fetching. The selected column must contain metabolite identifiers (e.g. HMBD, KEGG, ChEBI, etc)}

\item{out_col}{new column name for rowData to output pathway information to}

\item{pwdb_species}{name of the species the data was measured in. Use pathwayDatabases() after loading graphite to see a full list of databases and species}

\item{pwdb_name}{name of the pathway database to use. Use pathwayDatabases() after loading graphite to see a full list of databases and species}

\item{n_cores}{number of cores to use for parallelizaion (used in convertIdentifiers)}

\item{export_raw_db}{# OPTIONAL. Export the pathway database to a directory. Must be a string containing the path name with a .xlsx extension.}
}
\value{
rowData: new pathway annotation column added

$result$pathways: a dataframe of pathway information
}
\description{
Adds custom pathways to the already existing SummarizedExperiment
 data structure using the graphite package.
}
\examples{
# annotate metabolites using the humancyc database from the graphite package
\dontrun{... \%>\%
    mt_add_pathways(in_col = "KEGG",
                    out_col = "humancyc_db",
                    pwdb_species = "hsapiens",
                    pwdb_name = "humancyc",
                    n_cores = 5) \%>\%
...}

}
\author{
Parviz Gomari
}
