% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_pathways_xls.R
\name{mt_anno_pathways_xls}
\alias{mt_anno_pathways_xls}
\title{Add pathway information}
\usage{
mt_anno_pathways_xls(
  D,
  in_col,
  out_col,
  file,
  sheet,
  met_id,
  pw_id,
  pw_name,
  raw_db_outfile
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{in_col}{rowData column to use for pathway fetching. The selected column must contain metabolite
identifiers (e.g. HMBD, KEGG, ChEBI, etc).}

\item{out_col}{New column name for rowData to output pathway information to.}

\item{file}{Path where the pathway annotation flat file is stored.}

\item{sheet}{Sheet name or number to read in flat file.}

\item{met_id}{Name of flat file column containing metabolite IDs.}

\item{pw_id}{Name of flat file colname containing pathway IDs.}

\item{pw_name}{Name of flat file colname containing pathway names.}

\item{raw_db_outfile}{OPTIONAL. Name of file to export the pathway database to. Must be a string containing the path name with a
.xlsx extension.}
}
\value{
rowData: New pathway annotation column added.

$results$pathways: A dataframe of pathway information.
}
\description{
Adds custom pathways to the already existing SummarizedExperiment data structure using a flat file.
NOTE: The flat file should be an Excel file, containing 3 columns:
\itemize{
  \item{met_id} contains metabolite IDs
  \item{pw_id} contains pathway IDs
  \item{pw_name} contains pathway names}
The pathway columns pw_id and pw_name should be in long format, meaning if a metabolite is assigned to multiple pathways,
each metabolite pathway pair should appear in a separate row.
}
\examples{
\dontrun{# annotate metabolites using the SMP column of the pathway database flat file
... \%>\%
      mt_anno_pathways_xls(in_col = "HMDb_ID",
                           out_col = "janpw",
                           file = codes.makepath("packages/metabotools_external/hmdb/hmdb_preprocessed_4.0.csv"),
                           met_id = "HMDB_id",
                           pw_id = "SMP",
                           pw_name = "pathway_name") \%>\%
...}

}
\author{
PG
}
