% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_anno_xls.R
\name{mt_files_anno_xls}
\alias{mt_files_anno_xls}
\title{Load annotations from Excel file.}
\usage{
mt_files_anno_xls(
  D,
  file,
  sheet,
  anno_type,
  anno_ID,
  data_ID = anno_ID,
  no_map_err = F,
  col_names = NA
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{file}{input Excel file}

\item{sheet}{name or number of sheet}

\item{anno_type}{"samples" (colData) or "metabolites" (rowData) to add annotation columns to}

\item{anno_ID}{column in annotation file that contains ID information for mapping}

\item{data_ID}{column in existing annotation data frame that contains ID information for mapping}

\item{no_map_err}{throw error (T) or warning (F) if something does not map. default: F}

\item{col_names}{take column from new annotation file and overwite colnames (i.e. sample names) of SE; default: NA}
}
\value{
rowData or colData: new annotation columns added
}
\description{
If annotation fields are already existing, this function will fill up any NAs with the values from the new file. Existing values are not overwritten/
}
\details{
Loads annotations and merges them into current SummarizedExperiment.
 Performs "left-joins", i.e. leaves the original SE unchanged and just adds information where it can be mapped.
 Can load annotations for both metabolites (rowData) and samples (colData).
}
\examples{
# Load data, two sheets with sample annotations, and one sheet with metabolite annotations from the same file
\dontrun{D <-
  # load raw data
  mt_files_data_xls(file=file, sheet="data", samplesInRows=T, ID="SAMPLE_NAME") \%>\%
  # sample annotations from metabolomics run
  mt_files_anno_xls(file=file, sheet="sampleinfo", anno_type="samples", anno_ID = "SAMPLE_NAME") \%>\%
  # sample annotations from clinical table
  mt_files_anno_xls(file=file, sheet="clinicals", anno_type="samples", anno_ID="SAMPLE_NAME") \%>\%
  # metabolite annotations`
  mt_files_anno_xls(file=file, sheet="metinfo", anno_type="metabolites", anno_ID="BIOCHEMICAL", data_ID = "name") \%>\%
  ...}

}
\author{
JK
}
