% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_checksum.R
\name{mt_files_checksum}
\alias{mt_files_checksum}
\title{Validate MD5 checksum of file.}
\usage{
mt_files_checksum(D, file, checksum)
}
\arguments{
\item{D}{SummarizedExperiment for pass-through (missing if first step in pipeline)}

\item{file}{File path... does not necessarily have to be the same as the dataset loaded in the pipeline}

\item{checksum}{Checksum to test for.}
}
\value{
Does not change SummarizedExperiment, only adds a log message.
}
\description{
Validate MD5 checksum of file.
}
\examples{
# first call, to get the checksum (will crash, deliberately)
\dontrun{... \%>\% mt_files_checksum(file="input.xlsx", checksum="") \%>\% ...}

# copy-paste the correct ('actual') checksum from the error message into the call:
\dontrun{... \%>\% mt_files_checksum(file="input.xlsx", checksum="688048bd1eb9c771be0eb49548a6f947") \%>\% ...}

}
\author{
JK
}
