% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_data_xls.R
\name{mt_files_data_xls}
\alias{mt_files_data_xls}
\title{Load data matrix from Excel file.}
\usage{
mt_files_data_xls(D, file, sheet, samples_in_rows = T, ID_col, zero_to_NA = F)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input (missing if first step in pipeline)}

\item{file}{input Excel file}

\item{sheet}{name or number of sheet}

\item{samples_in_rows}{read samples as rows (T) or as columns (F). default: T (samples as rows)}

\item{ID_col}{if samples_in_rows==T -> name of sample ID column... must be exactly one
if samples_in_rows==F -> name of metabolite name column... must be exactly one}

\item{zero_to_NA}{replace zeros by NAs? (default: F)}
}
\value{
Produces an initial SummarizedExperiment, with assay, colData, rowData, and metadata with first entry
}
\description{
Loads numerical data matrix from Excel sheet.
Can handle files with samples in rows or columns.
}
\details{
Default name for entity in columns will be "name".
}
\examples{
# Load data, two sheets with sample annotations, and one sheet with metabolite annotations from the same file
\dontrun{D <-
  # load raw data
  mt_load_files_data_xls(file=file, sheet="data", samples_in_rows=T, ID_col="SAMPLE_NAME") \%>\%
  # sample annotations from metabolomics run
  mt_files_anno_xls(file=file, sheet="sampleinfo", annosfor="samples", IDanno = "SAMPLE_NAME") \%>\%
  # sample annotations from clinical table
  mt_files_anno_xls(file=file, sheet="clinicals", annosfor="samples", IDanno="SAMPLE_NAME") \%>\%
  # metabolite annotations`
  mt_files_anno_xls(file=file, sheet="metinfo", annosfor="metabolites", IDanno="BIOCHEMICAL", IDdata = "name") \%>\%
  ...}

}
\author{
JK
}
