% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_load_metabolon.R
\name{mt_files_load_metabolon}
\alias{mt_files_load_metabolon}
\title{Load Metabolon-format data.}
\usage{
mt_files_load_metabolon(file, sheet, copynansheet = "")
}
\arguments{
\item{file}{input Excel file}

\item{sheet}{name of sheet}

\item{copynansheet}{if given, which sheet to copy the NA pattern from}
}
\value{
Produces an initial SummarizedExperiment, with assay, colData, rowData, and metadata with first entry
}
\description{
Loads data from a Metabolon-format Excel file. Needs to be in the original "Client Data Table" format that they deliver.
}
\examples{
\dontrun{D <- 
  # load data
  mt_files_load_metabolon(codes.makepath("Mt/sampledata.xlsx"), "OrigScale") \%>\%
  ...}

}
\author{
JK
}
