% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_load_metabolon_lipidomics.R
\name{mt_files_load_metabolon_lipidomics}
\alias{mt_files_load_metabolon_lipidomics}
\title{Load Metabolon-format lipidomics data.}
\usage{
mt_files_load_metabolon_lipidomics(file, sheet_list)
}
\arguments{
\item{file}{input Excel file}

\item{sheet_list}{list with the sheet names to read}

\item{copynansheet}{if given, which sheet to copy the NA pattern from}
}
\value{
Produces an initial SummarizedExperiment, with assay, colData, rowData, and metadata with first entry
}
\description{
Loads lipidomics data from a Metabolon-format Excel file. Needs to be in the original "Client Data Table" format that they deliver.
}
\examples{
D <- 
  # load data
  mt_files_load_metabolon_lipidomics(codes.makepath("Mt/sampledata.xlsx"), sheet_list = c("Species Concentrations","Fatty Acid Concentrations")) \%>\%
  ...

}
\author{
EB
}
