% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_load_nightingale.R
\name{mt_files_load_nightingale}
\alias{mt_files_load_nightingale}
\title{Load Nightingale-format data.}
\usage{
mt_files_load_nightingale(
  excel_file,
  raw_sheet,
  met_sheet,
  sample_id = "Sample id"
)
}
\arguments{
\item{excel_file}{input Excel file}

\item{raw_sheet}{name of sheet with raw data}

\item{met_sheet}{name of sheet with biomarker information}

\item{sample_id}{optional, name of the column with sample id}
}
\value{
Produces an initial SummarizedExperiment, with assay, colData, rowData, and metadata with first entry
}
\description{
Loads data from a Nightingale format Excel file.
}
\examples{
\dontrun{D <-
  # load data
  mt_files_load_nightingale(codes.makepath("Mt/sampledata_night.xlsx"), "Results") \%>\%
  ...}

}
\author{
RB
}
