% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_load_olink.R
\name{mt_files_load_olink}
\alias{mt_files_load_olink}
\title{Load Olink-format data.}
\usage{
mt_files_load_olink(D, file)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input (missing if first step in pipeline)}

\item{filename}{filename of NPX file exported from NPX manger.}
}
\value{
Produces an initial SummarizedExperiment, with assay (note: 2^NPX), colData, rowData, and metadata with first entry
}
\description{
Loads data from an Olink-format Excel file.
Uses Olink's R code read_NPX taken from https://github.com/Olink-Proteomics/OlinkRPackage/tree/master/OlinkAnalyze
}
\details{
In case the Olink file is not in XLSX format, but CSV or TSV text format:
you may need to remove all Ctrl data columns
save file as xlsx (make sure to rename SEPT9 back from Excel date)
don't keep the Excel file open while running R - this throws a file access denied error
}
\examples{
D <-
  # load data
  mt_load_files_olink(codes.makepath("mt/olink_sampledata.xlsx")) \%>\%
  ...

}
\author{
KS
}
