% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_load_UCD.R
\name{mt_files_load_UCD}
\alias{mt_files_load_UCD}
\title{Load UC Davis-format data.}
\usage{
mt_files_load_UCD(D, file, sheet, zero_to_NA = T, gen_valid_varnames = F)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input (missing if first step in pipeline)}

\item{file}{input Excel file in UCD format}

\item{sheet}{name or number of sheet}

\item{zero_to_NA}{replace zeros by NAs? (default: T)}

\item{gen_valid_varnames}{enforce valid R variable names?}
}
\value{
Produces an initial SummarizedExperiment, with assay, colData, rowData, and metadata with first entry
}
\description{
Loads data from a UC Davis-format Excel file.
}
\examples{
\dontrun{D <-
  # load data
  mt_load_files_UCD(codes.makepath("Mt/sampledata.xlsx"), "OrigScale") \%>\%
  ...}

}
\author{
JK
}
