% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_write_SE.R
\name{mt_files_write_SE}
\alias{mt_files_write_SE}
\title{Save SummarizedExperiment to file}
\usage{
mt_files_write_SE(D, file)
}
\arguments{
\item{D}{SummarizedExperiment input}

\item{file}{File to write to.}
}
\description{
Simple helper function to save SE to a file. Object name will be \code{D}
}
\examples{
\dontrun{#
... \%>\% mt_files_write_SE(file="out.RDS") \%>\% ...
...}

}
\author{
JK
}
