% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_write_stats.R
\name{mt_files_write_stats}
\alias{mt_files_write_stats}
\title{Export statistical results from pipeline into Excel file.}
\usage{
mt_files_write_stats(D, file, compnames = NULL, sort.by.p = F)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{file}{output filename to write to}

\item{compnames}{Names of one or more statistical comparison to be written out. If NULL, will export all.}

\item{sort.by.p}{Automatically sort by p-values? (default: False)}
}
\value{
Does not change the SummarizedExperiment.
}
\description{
Writes out the statistics data frame with p-values, adjusted p-values, fold changes, test statistics etc. into an Excel sheet.
}
\examples{
\dontrun{# Write out all results
... \%>\% mt_files_write_stats(file="results.xlsx") \%>\%
# Write out specific result]
... \%>\% mt_files_write_stats(file="results.xlsx", compnames="comp1") \%>\%}

}
\author{
JK
}
