% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_flag_logged.R
\name{mt_flag_logged}
\alias{mt_flag_logged}
\title{Actively flags data as logged.}
\usage{
mt_flag_logged(D)
}
\arguments{
\item{D}{SummarizedExperiment input}
}
\value{
unchanged SE
}
\description{
# Required when loading already-logged data, to signal to other functions (e.g. fold change calculation) that this is logged data.
 Technically, function does not do anything, but leaves its call footprint in the pipeline to be found by others.
}
\examples{
\dontrun{# simply insert into pipeline
... \%>\% mt_flag_logged() \%>\%
...}

}
\author{
JK
}
