% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_data_xls.R
\name{mt_load_data_xls}
\alias{mt_load_data_xls}
\title{Load data matrix from Excel file}
\usage{
mt_load_data_xls(D, file, sheet, samples_in_rows = T, id_col, zero_to_na = F)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. Missing if first step in pipeline.}

\item{file}{Name of input Excel file.}

\item{sheet}{Name or number of sheet.}

\item{samples_in_rows}{Read samples as rows (T) or as columns (F). Default: T (samples as rows).}

\item{id_col}{If samples_in_rows==T -> name of sample ID column. If samples_in_rows==F -> name of metabolite name
column. Must be exactly one.}

\item{zero_to_na}{Replace zeros by NAs? Default: F.}
}
\value{
If first step in pipeline, creates \code{SummarizedExperiment} object. Populates empty assay, colData, and rowData data frames.
}
\description{
Loads numerical data matrix from Excel sheet. Can handle files with samples in rows or columns. Default name for entity in
columns will be "name".
}
\examples{
# Load data, two sheets with sample annotations, and one sheet with metabolite annotations from the same file
\dontrun{D <-
  # load raw data
  mt_load_files_data_xls(file=file, sheet="data", samples_in_rows=T, id_col="SAMPLE_NAME") \%>\%
  # sample annotations from metabolomics run
  mt_files_anno_xls(file=file, sheet="sampleinfo", annosfor="samples", IDanno = "SAMPLE_NAME") \%>\%
  # sample annotations from clinical table
  mt_files_anno_xls(file=file, sheet="clinicals", annosfor="samples", IDanno="SAMPLE_NAME") \%>\%
  # metabolite annotations`
  mt_files_anno_xls(file=file, sheet="metinfo", annosfor="metabolites", IDanno="BIOCHEMICAL", IDdata = "name") \%>\%
  ...}

}
\author{
JK
}
