% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_files_WCM.R
\name{mt_load_files_WCM}
\alias{mt_load_files_WCM}
\title{Load WCM core-format data.}
\usage{
mt_load_files_WCM(D, file, sheet, zero_to_NA = T)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input (missing if first step in pipeline)}

\item{file}{input Excel file}

\item{sheet}{name or number of sheet}

\item{zero_to_NA}{replace zeros by NAs? (default: T)}
}
\value{
Produces an initial SummarizedExperiment, with assay, colData, rowData, and metadata with first entry
}
\description{
Loads data from a WCM core format file (as they send it).
}
\examples{
\dontrun{D <-
  # load data in WCM format
  mt_load_files_WCM(
    file='DM369_Metabolites separated by nucleotides 12_6_18_altered.xlsx',
    sheet='peakHeight_metabolite_intensiti') \%>\%
  ...}

}
\author{
JK
}
