% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_load_metabolon_lipidomics.R
\name{mt_load_metabolon_lipidomics}
\alias{mt_load_metabolon_lipidomics}
\title{Load Metabolon-format lipidomics data}
\usage{
mt_load_metabolon_lipidomics(D, file, sheet_list)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. Missing if first step in pipeline.}

\item{file}{Name of input Excel file.}

\item{sheet_list}{List with the sheet names to read.}
}
\value{
If first step in pipeline, creates \code{SummarizedExperiment} object. Populates empty assay, colData, and rowData data frames.
}
\description{
Loads lipidomics data from a Metabolon-format Excel file. Needs to be in the original "Client Data Table" format that they deliver.
}
\examples{
D <-
  # load data
  mt_load_files_metabolon_lipidomics(codes.makepath("Mt/sampledata.xlsx"), sheet_list = c("Species Concentrations","Fatty Acid Concentrations")) \%>\%
  ...

}
\author{
EB
}
