% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_logging_datasetinfo.R
\name{mt_logging_datasetinfo}
\alias{mt_logging_datasetinfo}
\title{Output information about dataset to log.}
\usage{
mt_logging_datasetinfo(D)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}
}
\value{
Nothing - leaves \code{D} unchanged
}
\description{
Leaves a log entry containing the number of samples, metabolites and annotation fields at the current stage of the pipeline.
}
\examples{
\dontrun{...  \%>\%
  mt_logging_datasetinfo() \%>\% ...}

}
\author{
JK
}
