% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_logging_statsinfo.R
\name{mt_logging_statsinfo}
\alias{mt_logging_statsinfo}
\title{Output information about statistical results.}
\usage{
mt_logging_statsinfo(D, stat_name, stat_filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{stat_name}{name of the statistical comparison}

\item{stat_filter}{filter formula to display number of results, e.g. p.adj<0.2}
}
\value{
Nothing - leaves \code{D} unchanged
}
\description{
Leaves a log entry containing the number of samples in each group that was used, and the number of hits according to a formula.
}
\examples{
\dontrun{...  \%>\%
  mt_logging_statsinfo(stat_name="comp1", stat_filter=p.adj<0.2) \%>\% ...}

}
\author{
JK
}
