% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_logging_tictoc.R
\name{mt_logging_toc}
\alias{mt_logging_toc}
\title{Timing functionality.}
\usage{
mt_logging_toc(D)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}
}
\value{
Nothing - leaves \code{D} unchanged
}
\description{
Call mt_logging_tic to start timing anywhere in pipeline, then mt_logging_toc to show the time elapsed in the status log.
}
\examples{
\dontrun{\%>\%
  mt_logging_tic() \%>\%
  ... \%>\% ... # pipeline steps
  ... \%>\% ... # pipeline steps
  ... \%>\% ... # pipeline steps
  mt_logging_toc()}

}
\author{
JK
}
