% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_applytoanno.R
\name{mt_modify_applytoanno}
\alias{mt_modify_applytoanno}
\title{Apply function to metabolite or sample annotation column.}
\usage{
mt_modify_applytoanno(D, anno_type, col_name, fun)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{anno_type}{"samples" or "metabolites"}

\item{col_name}{name of column to access}

\item{fun}{function to be applied}
}
\value{
changes the contents of a column either in colData or rowData
}
\description{
Works in the same way as lapply/sapply, with a function to be executed on each entry of a given metabolite or sample annotation.
}
\examples{
 \dontrun{... \%>\%
 # ensure factor for casecontrol variable
 mt_modify_applytoanno(
   anno_type='samples',
   col_name='casecontrol',
   fun=as.factor) \%>\%
 ...}

}
\author{
JK
}
