% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_averagemetabolite.R
\name{mt_modify_averagemetabolite}
\alias{mt_modify_averagemetabolite}
\title{Averages duplicate metabolites}
\usage{
mt_modify_averagemetabolite(D, avg_by)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{avg_by}{name of rowData column (metabolite annotation) by which duplicates can be identified}
}
\value{
D with duplicate metabolites combined
}
\description{
Averages values of metabolites (rows) with same values in specified rowData column.
}
\examples{
\dontrun{... \%>\% mt_modify_averagemetabolite(avg_by = 'name') \%>\% ...}

}
\author{
RB
}
