% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_averagesample.R
\name{mt_modify_averagesample}
\alias{mt_modify_averagesample}
\title{Averages duplicate samples.}
\usage{
mt_modify_averagesample(D, group_by)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{group_by}{name of colData column (sample annotation) by which duplicates can be identified}
}
\value{
D with duplicate samples combined
}
\description{
{ADD DESCRIPTION}
}
\examples{
\dontrun{... \%>\% mt_modify_averagesample(group_by = "RID") \%>\% ...}

}
\author{
Annalise Schweickart
}
