% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_avg_metabolite.R
\name{mt_modify_avg_metabolite}
\alias{mt_modify_avg_metabolite}
\title{Averages duplicate metabolites}
\usage{
mt_modify_avg_metabolite(D, group_col)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{group_col}{Name of rowData column (metabolite annotation) by which duplicates can be identified.}
}
\value{
D with duplicate metabolites combined
}
\description{
{ADD DESCRIPTION}
}
\examples{
\dontrun{... \%>\% mt_modify_avg_metabolite(group_col = 'name') \%>\% ...}

}
\author{
RB
}
