% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_avg_metabolites.R
\name{mt_modify_avg_metabolites}
\alias{mt_modify_avg_metabolites}
\title{Averages duplicate metabolites}
\usage{
mt_modify_avg_metabolites(D, group_col)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{group_col}{Name of rowData column (metabolite annotation) by which duplicates can be identified.}
}
\value{
assay: duplicate metabolites (rows) combined
}
\description{
Averages values of metabolites (rows) with same values in specified rowData column.
}
\examples{
\dontrun{... \%>\% mt_modify_avg_metabolites(group_col = 'name') \%>\% ...}

}
\author{
RB
}
