% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_avg_sample.R
\name{mt_modify_avg_sample}
\alias{mt_modify_avg_sample}
\title{Averages duplicate samples}
\usage{
mt_modify_avg_sample(D, group_col)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{group_col}{Name of colData column (sample annotation) by which duplicates can be identified.}
}
\value{
D with duplicate samples combined
}
\description{
{ADD DESCRIPTION}
}
\examples{
\dontrun{... \%>\% mt_modify_avg_sample(group_col = "RID") \%>\% ...}

}
\author{
AS
}
