% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_avgmetabolite.R
\name{mt_modify_avgmetabolite}
\alias{mt_modify_avgmetabolite}
\title{mt-modify_avgmetabolite}
\usage{
mt_modify_avgmetabolite(D, avg_by)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{avg_by}{name of rowData column (metabolite annotation) by which duplicates can be identified}
}
\value{
D with duplicate metabolites combined
}
\description{
Averages duplicate metabolites
}
\examples{
\dontrun{... \%>\% mt_modify_avgmetabolite(avg_by = 'name') \%>\% ...}

}
\author{
RB
}
