% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_cv.R
\name{mt_modify_cv}
\alias{mt_modify_cv}
\title{mt_modify cv}
\usage{
mt_modify_cv(D, qc_samples, col_lab)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{qc_samples}{Logical expression. Can use fields from \code{colData()}}

\item{col_lab}{Label for the new cv values to be stored in in colData}
}
\value{
rowData(D)[[col_lab]] contains column with CV score for each metabolite
}
\description{
Calculate CV score
}
\examples{
\dontrun{... \%>\% mt_modify_cv(qc_samples=="PQC", col_lab = "PQC_cv") \%>\% ...}

}
\author{
Annalise Schweickart
}
