% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_filter.R
\name{mt_modify_filter_metabolites}
\alias{mt_modify_filter_metabolites}
\title{Filter metabolites}
\usage{
mt_modify_filter_metabolites(D, filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{filter}{Logical expression. Can use columns from rowData.}
}
\value{
assay: Filtered metabolites (rows) removed.

rowData: Filtered metabolites removed.
}
\description{
Filters metabolites according to an expression. Expression can access entries of rowData.
}
\examples{
\dontrun{... \%>\% mt_modify_filter_metabolites(SUPER_PATHWAY=="Nucleotide") \%>\% ...}

}
\author{
JK
}
