% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_icc.R
\name{mt_modify_icc}
\alias{mt_modify_icc}
\title{mt_modify_icc}
\usage{
mt_modify_icc(D, qc_samples, col_lab, id_col, icc_lmer = F)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{qc_samples}{Logical expression. Can use fields from \code{colData()}}

\item{col_lab}{Label for the new icc values to be stored in in rowData}

\item{id_col}{it is the name of the column in colData containing ids qc_samples}

\item{icc_lmer}{Optional flag to set lmer flag in ICC function default is FALSE}
}
\value{
rowData(D)[[col_lab]] contains column with ICC score for each metabolite
}
\description{
Calculate ICC - intraclass correlation
}
\examples{
\dontrun{... \%>\% mt_modify_icc(qc_samples=="PQC", col_lab = "PQC_cv", id_col='RID') \%>\% ...}

}
\author{
RB
}
