% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_missingness.R
\name{mt_modify_missingness}
\alias{mt_modify_missingness}
\title{Add Missingness Column to rowData / colData}
\usage{
mt_modify_missingness(D, add_mets = T, add_samples = T)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{add_mets}{add metabolite missingness? default: T}

\item{add_samples}{add sample missingness? default: T}
}
\value{
if add_mets == T: new rowData column

if add_samples == T: new colData column
}
\description{
Adds a missingness column for rowData (metabolites) and / or colData (samples).
}
\examples{
# example of how to run function
\dontrun{}

}
\author{
KC
}
