% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_ratios.R
\name{mt_modify_ratios}
\alias{mt_modify_ratios}
\title{Generate metabolite ratios.}
\usage{
mt_modify_ratios(D)
}
\arguments{
\item{D}{SummarizedExperiment object}
}
\value{
SummarizedExperiment containing pairwise ratios from all variables of input
}
\description{
Transforms the dataset into a new dataset where each 'metabolite' represents a ratio of two metabolites. Warning: For a dataset
with originally p metabolites, this will result in p*(p-1) new variables. (e.g. 500 metabolits becomes 249500 ratios).
}
\details{
mt_post_pgain provides a special operation on a ratio data matrix for better interpretation of the resulting p-values.
}
\examples{
\dontrun{# Transform dataset to ratios
... \%>\%  mt_modify_ratios() \%>\% ... # proceed with statistical analysis
}

}
\author{
Jonas Zierer, JK
}
