% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_replaceSE.R
\name{mt_modify_replaceSE}
\alias{mt_modify_replaceSE}
\title{Replace SummarizedExperiment with another one.}
\usage{
mt_modify_replaceSE(D, Drepl)
}
\arguments{
\item{D}{SummarizedExperiment object}

\item{Drepl}{replacement SummarizedExperiment object where assay(), colData() and rowData() will be copied from.}
}
\value{
SummarizedExperiment containing assay(), colData() and rowData() of Drepl, but metadata of the original pipeline object
}
\description{
This function replaces the assay(), colData() and rowData() of the current pipeline object with those of another one. 
The metadata of the original pipeline object is retained.
}
\details{
The function provides a possibility for "pipeline restarts" after operations that entirely change the data matrix, such
as mt_modify_ratios and mt_modify_aggPW.
}
\examples{
\dontrun{# Transform dataset to ratios
... \%>\%  mt_modify_replaceSE(Drepl = D1) \%>\% ... # replace with an earlier SE
}

}
\author{
JK
}
