% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_other_settings.R
\name{mt_other_settings}
\alias{mt_other_settings}
\title{Sets and outputs global pipeline settings}
\usage{
mt_other_settings(D, settings)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input. Missing if first step in pipeline.}

\item{settings}{List of settings.}
}
\value{
If first step in pipeline, creates \code{SummarizedExperiment}. Otherwise, does not change the
   \code{SummarizedExperiment} object.
}
\description{
Call without parameters to show current settings. Call with list of parameters to set settings. Will crash for invalid setting names.
}
\examples{
\dontrun{...  \%>\%
  mt_other_settings(list(ggplot_fix=F)) \%>\% ... # deactivate ggplot fixing
 }

}
\author{
JK
}
