% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_other_strip_results.R
\name{mt_other_strip_results}
\alias{mt_other_strip_results}
\title{Remove metadata results from SummarizedExperiment}
\usage{
mt_other_strip_results(D, strip = "all")
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{strip}{What to remove. Has to be "all" or "plots". Default: "all".}
}
\description{
strip=="all" will only leave a single result in the object, the one from this function.

strip=="plots" will set all ggplot objects to NULL, still allowing HTML reports to be generated.
}
\details{
Removes either the entire result list or all plots from the metadata of a SummarizedExperiment.
Can be used to obtain a more lightweight object for further processing.
}
\examples{
\dontrun{# at the end of a pipeline
... \%>\% mt_other_strip_results(strip="plots")
...}

}
\author{
JK
}
