% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_box.R
\name{mt_plots_box}
\alias{mt_plots_box}
\title{Plot Box Plots}
\usage{
mt_plots_box(
  D,
  x = "x",
  stat_name,
  correct_confounder,
  metab_filter = p.value < 0.05,
  metab_sort = p.value,
  annotation = "{sprintf('P-value: \%.1e', p.value)}",
  text_size = 3.88,
  jitter = T,
  rows,
  cols,
  restrict_to_used_samples = T,
  full_info = F,
  manual_ylabel = NULL,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{x}{What phenotype (from colData) should be used on x axis. Default: "x".}

\item{stat_name}{Index of the entry in metadata(D)$results that contains statistic object.}

\item{correct_confounder}{Confounders to adjust for before plotting, formula notation.}

\item{metab_filter}{If given, filter will be applied to data and remaining variables will be labeled in plot. Default p.value<0.05.}

\item{metab_sort}{If given, arrange will be applied to data variables will be sorted. Default: p.value.}

\item{text_size}{Text size of the annotations. Default: 3.88.}

\item{jitter}{Whether to add jitter to boxplot. Default T.}

\item{rows}{Number rows of boxplots in $result.}

\item{cols}{Number columns of boxplots in $result.}

\item{restrict_to_used_samples}{Whether to filter to the samples that were used in the statistical test. Default: T.}

\item{full_info}{Add full information of all sample annotations and statistics results to plottable data.frame? makes
plotting more flexible but can render SE objects huge. Default: F.}

\item{manual_ylabel}{Manual ylabel. Default: NULL.}

\item{ggadd}{Further elements / functions to add (+) to the ggplot object.}

\item{...}{Additional expression directly passed to aes() of ggplot, can refer to colData.}
}
\value{
$result$output: plot, boxplot
}
\description{
Creates one boxplot plot per metabolite based on given sample annotations.
}
\examples{
\dontrun{# box plots as overview of results with a result already in 'comp'
# color by "Group" variable in colData
mt_plots_box(x                  = Group,
                 stat_name           = "comp",
                 correct_confounder = ~BATCH_MOCK,
                 metab_filter       = p.value<0.01,
                 metab_sort         = p.value,
                 annotation         = "{sprintf('P-value: \%.1e', p.value)}\nStatistic: {sprintf('\%.2f', statistic)}",
                 rows               = 2,
                 cols               = 2) \%>\%
                 ...
 }

}
\author{
JZ
}
