% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_boxplot_scatter.R
\name{mt_plots_boxplot_scatter}
\alias{mt_plots_boxplot_scatter}
\title{Plot Box Plots or Scatter Plots}
\usage{
mt_plots_boxplot_scatter(
  D,
  x = "x",
  stat_name,
  plot_type,
  correct_confounder,
  metab_filter = p.value < 0.05,
  metab_sort = p.value,
  annotation = "{sprintf('P-value: \%.1e', p.value)}",
  full_info = F,
  text_size = 3.88,
  jitter = "beeswarm",
  restrict_to_used_samples = T,
  manual_ylab = NULL,
  fitline = T,
  fitline_se = T,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{x}{what phenotype (from colData(D)) should be used on x axis, default "x"}

\item{stat_name}{index of the entry in metadata(D)$results that contains statistic object}

\item{plot_type}{"box" or "scatter"}

\item{correct_confounder}{confounders to adjust for before plotting, formula notation}

\item{metab_filter}{if given, filter will be applied to data and remaining variables will be labelled in plot, default p.value<0.05}

\item{metab_sort}{if given, arrange will be applied to data variables will be sorted, default p.value}

\item{full_info}{add full information of all sample annotations and statistics results to plottable data.frame? makes plotting more flexible but can render SE objects huge. default: F}

\item{text_size}{text size of the annotations}

\item{jitter}{add geom_jitter ("jitter") or geom_beeswarm ("beeswarm") to boxplot, exclude if NULL;  default "beeswarm"}

\item{restrict_to_used_samples}{whether to filter to the samples that were used in the statistical test, default: T}

\item{manual_ylab}{manual ylabel (default: none)}

\item{fitline}{add fit line? (default: T)}

\item{fitline_se}{add standard error range? (default: T)}

\item{ggadd}{further elements/functions to add (+) to the ggplot object}

\item{...}{additional expression directly passed to aes() of ggplot, can refer to colData}
}
\value{
if plot_type = "box", $result: plot, box plot

if plot_type = "scatter", $result: plot, scatter plot
}
\description{
Creates one box plot or  scatter plot per metabolite based on given sample annotations
}
\author{
JZ, KC
}
