% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_glmnet_roc_auc.R
\name{mt_plots_glmnet_roc_auc}
\alias{mt_plots_glmnet_roc_auc}
\title{Plot Evaluation Measures for glmnet (binary) Predictions}
\usage{
mt_plots_glmnet_roc_auc(
  D,
  result_name,
  plot_measures = c("spec", "sens", "ppv", "f1", "acc"),
  cutoff,
  pos_class
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{result_name}{name under which the glmnet results are stored, must be unique to all other glmnet results}

\item{plot_measures}{a vector of measures to include; can be one or more of the following: spec, sens, f1, acc, and/or ppv; default c("spec", "sens", "ppv")}

\item{cutoff}{threshold cutoff for prediction probability class assignment; function will select threshold closest to user provided value}

\item{pos_class}{which label to set as positive class}
}
\value{
results$ouput: list of four or more plots
}
\description{
This function creates four plots: (1) ROC curve, (2) evaluation measures across thresholds (folds ignored), (3) AUC scores per fold,
(4) evaluation measures per fold (if cutoff provided, single values; else across thresholds).
}
\examples{
# example of how to run function
\dontrun{D \%<>\% mt_plots_glmnet_roc_auc(result_name = "glmnet result", plot_measures = c("spec", "sens") cutoff = 0.75)}

}
\author{
KC
}
