% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_missingness.R
\name{mt_plots_missingness}
\alias{mt_plots_missingness}
\title{Visualize missing value structure of dataset}
\usage{
mt_plots_missingness(
  D,
  met_max = NA,
  samp_max = NA,
  plot_mets = T,
  plot_samples = F,
  sec_axis_mets = F,
  sec_axis_samples = F,
  sample_labels = NA,
  plot_data = T
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{met_max}{Which metabolite missingness to mark on the y axis. Default: NA (no line).}

\item{samp_max}{Which sample missingness to mark on the y axis. Default: NA (no line).}

\item{plot_mets}{Show metabolite missingness plot? Default: T.}

\item{plot_samples}{Show sample missingness plot? Default: F.}

\item{sec_axis_mets}{Add second axis to metabolite missingness plot? Default: F.}

\item{sec_axis_samples}{Add second axis to sample missingness plot? Default: F.}

\item{sample_labels}{Which column from colData to use as sample labels? Default: NA.}

\item{plot_data}{Show entire data missingness plot? Default: T.}
}
\value{
$result$output: plot, two plots
}
\description{

}
\examples{
\dontrun{\%>\%  mt_plots_missingness() \%>\% # without horizontal line
\%>\%  mt_plots_missingness(met_max=0.5) \%>\% # with horizontal line at 50\%
}

}
\author{
JK
}
