% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_net.R
\name{mt_plots_net}
\alias{mt_plots_net}
\title{mt_plots_net}
\usage{
mt_plots_net(
  D,
  stat_name,
  corr_filter = p.value < 0.05,
  node_coloring,
  save.html,
  height = 500
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{stat_name}{name of the test to take correlations from}

\item{corr_filter}{filter for correlation values to plot}

\item{node_coloring}{name of the test to use for node coloring}

\item{save.html}{filename of visnetwork html file. If empty, no html saved}

\item{height}{optional size in pixel of the plotting window size. Default 500px.}
}
\value{
assay: not altered

$result: network ggplot + visnetwork plot
}
\description{
Creates a network object
}
\examples{
\dontrun{#' # in the context of a SE pipeline
... \%>\% mt_plots_net(statsname = "xxx") \%>\% ...    # standard call
... \%>\% mt_plots_net(statsname = "xxx", corr_filter = p.adj < 0.5, node_coloring="Li's", save.html="Network.html", height=800) \%>\% ...    # filters only significant correlations and colors the nodes according to the results in the indicated test, saves visnetwork to file
}

}
\author{
EB
}
