% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_pathview.R
\name{mt_plots_pathview}
\alias{mt_plots_pathview}
\title{mt_plots_pathview}
\usage{
mt_plots_pathview(
  D,
  gene.id = NULL,
  gene.data = NULL,
  gene.idtype = "entrez",
  met.id = NULL,
  cpd.data = NULL,
  cpd.idtype = "kegg",
  stat_name,
  metab.filter,
  color.scale,
  color.range,
  show.only.filtered = FALSE,
  low = list(gene = "green", cpd = "yellow"),
  mid = list(gene = "gray", cpd = "gray"),
  high = list(gene = "red", cpd = "blue"),
  pathway.id,
  n.pathways,
  path.database = "./Pathview_database",
  path.output = "./Pathview_output",
  same.layer = TRUE,
  out.suffix = "pathview",
  add.pwname.suffix = FALSE,
  db_path = data.makepath("MT_precalc/pathview/KeggPathways.Rds"),
  species = "hsa",
  gene.annotpkg = NULL,
  min.nnodes = 3,
  kegg.native = TRUE,
  map.null = TRUE,
  expand.node = FALSE,
  split.group = FALSE,
  map.symbol = TRUE,
  map.cpdname = TRUE,
  node.sum = "sum",
  discrete = list(gene = FALSE, cpd = FALSE),
  limit = list(gene = 1, cpd = 1),
  bins = list(gene = 10, cpd = 10),
  both.dirs = list(gene = TRUE, cpd = TRUE),
  trans.fun = list(gene = NULL, cpd = NULL),
  na.col = "transparent"
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{gene.id}{string name of the rowData column containing the gene identifiers.}

\item{gene.data}{either vector (single sample) or a matrix-like data (multiple sample). Vector should be numeric with gene IDs as names or it may also be character of gene IDs. Character vector is treated as discrete or count data. Matrix-like data structure has genes as rows and samples as columns. Row names should be gene IDs. Here gene ID is a generic concepts, including multiple types of gene, transcript and protein uniquely mappable to KEGG gene IDs. KEGG ortholog IDs are also treated as gene IDs as to handle metagenomic data. Check details for mappable ID types. Default gene.data=NULL.}

\item{gene.idtype}{character, ID type used for the gene.data, case insensitive. Default gene.idtype="entrez", i.e. Entrez Gene, which are the primary KEGG gene ID for many common model organisms. For other species, gene.idtype should be set to "KEGG" as KEGG use other types of gene IDs. For the common model organisms (to check the list, do: data(bods); bods), you may also specify other types of valid IDs. To check the ID list, do: data(gene.idtype.list); gene.idtype.list.}

\item{met.id}{string name of the rowData column containing the metabolite identifiers}

\item{cpd.data}{the same as gene.data, excpet named with IDs mappable to KEGG compound IDs. Over 20 types of IDs included in CHEMBL database can be used here. Check details for mappable ID types. Default cpd.data=NULL. Note that gene.data and cpd.data can't be NULL simultaneously.}

\item{cpd.idtype}{character, ID type used for the cpd.data. Currently only works with "kegg".}

\item{stat_name}{name of the statistics object to apply metab.filter to}

\item{metab.filter}{if given, filter will be applied to data and only variables satisfying the condition will be included}

\item{color.scale}{if given, this will be used to map colors to a continuous scale}

\item{color.range}{numeric (positive), if given, indicates the color range (-color.range, +color.range). If missing, color.range will be determined internally.}

\item{show.only.filtered}{logical, if TRUE generate pathway list only based on filtered variables, otherwise pathways will be generated based on all variables.}

\item{low, mid, high}{each is a list of two colors with "gene" and "cpd" as the names. This argument specifies the color spectra to code gene.data and cpd.data. Default spectra (low-mid-high) "green"-"gray"-"red" and "yellow"-"gray"-"blue" are used for gene.data and cpd.data respectively. The values for 'low, mid, high' can be given as color names ('red'), plot color index (2=red), and HTML-style RGB, ("\#FF0000"=red).}

\item{pathway.id}{character vector, the KEGG pathway ID(s), usually 5 digit, may also include the 3 letter KEGG species code. If missing, the function will find all KEGG pathway annotations for the given KEGG identifiers.}

\item{n.pathways}{(optional) number of pathways to output. Most populated pathway will be plotted first.}

\item{path.database}{character, the directory path of KEGG pathway data file (.xml) and image file (.png). If the path does not exist, the function will create it. Default path.database = "./Pathview_database" (subfolder in the current working directory).}

\item{path.output}{character, the directory path of the function output files. If the path does not exist, the function will create it. Default path.output ="./Pathview_output" (subfolder in the current working directory).}

\item{same.layer}{logical, controls if node colors are to be plotted in the same layer as the pathway graph. If FALSE, output generation will be faster, but output plots will be larger in size.}

\item{out.suffix}{character, the suffix to be added after the pathway name as part of the output graph file. Default out.suffix="pathview".}

\item{add.pwname.suffix}{logical, if TRUE will add the pathway name to the output filename. If FALSE, will use what stored in out.suffix for all files. Default add.pwname.to.filename=FALSE.}

\item{db_path}{path to the pathway database to read annotations from.}

\item{\dots}{see \code{pathview::pathview} for pathview arguments}
}
\value{
$result: pathview images
}
\description{
Creates KEGG pathways visualization
}
\examples{
\dontrun{# plot all pathways with at least one significant metabolite from the statistical comparison "comp" in them
mt_plots_pathview(D = D,
                  met.id="KEGG_mapped",
                  stat_name = "comp",
                  color.scale = -sign(fc)*log10(p.adj),
                  color.range = -log10(0.01),
                  metab.filter = p.adj < 0.05,
                  show.only.filtered = TRUE,
                  path.database = "./Pathview_database",
                  path.output = "./results/pathview",
                  same.layer = F,
                  add.pwname.suffix = T
                  ) \%>\%
                  ...}

}
\author{
Elisa Benedetti
}
