% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_PCA.R
\name{mt_plots_PCA}
\alias{mt_plots_PCA}
\title{2D PCA of samples.}
\usage{
mt_plots_PCA(
  D,
  title = "PCA",
  scaledata = F,
  PCa = 1,
  PCb = 2,
  show = "scores",
  labelby = "",
  textrepel = T,
  ellipse = NA,
  expvarplot = F,
  store.matrices = F,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{title}{Title of the plot, default: "PCA"}

\item{scaledata}{scale data before plotting? (mean 0, std 1), default: FALSE}

\item{PCa}{first PC to plot, default is 1 (PC1)}

\item{PCb}{second PC to plot, default is 2 (PC2)}

\item{show}{show 'scores' or 'loadings'}

\item{labelby}{field to label. default: none}

\item{textrepel}{try to avoid all text overlaps when labeling? default:T}

\item{ellipse}{confidence interval for ellipse. default: none (no ellipse)}

\item{expvarplot}{add explained variance plot? default: F}

\item{store.matrices}{store scores and loadings matrices in result structure? default: F}

\item{ggadd}{further elements/functions to add (+) to the ggplot object}

\item{...}{# additional expression directly passed to aes() of ggplot, can refer to colData}
}
\value{
result output: plot(s)

result output2: scores and loadings matrix
}
\description{
Can be colored by any variable in colData.
}
\examples{
\dontrun{## PCA on scaledata, color and shape by "Group" variable in colData
... $>$ mt_plots_PCA(scaledata=T, color=Group, shape=Group, title="PCA - scaled data") \%>\% ...
## PCA scores plot on non-scaled data, with ellipse and extra explained variance plot, and two ggadds (white background and centering of title)
mt_plots_PCA(title="PCA scores", show = 'scores', scaledata=F, PCa=1, PCb=2, ellipse=0.95, expvarplot=T, ggadd = theme_bw() + theme(plot.title=element_text(hjust=0.5)))
}

}
\author{
JK, BGP
}
