% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_pheatmap_multStats.R
\name{mt_plots_pheatmap_multstats}
\alias{mt_plots_pheatmap_multstats}
\title{Heatmap of Multiple Statistical Objects}
\usage{
mt_plots_pheatmap_multstats(
  D,
  stat_list = NA,
  color_formula = sign(statistic) * -log10(p.adj),
  met_anno,
  signif_expr,
  signif_mark = "•",
  mark_size = 20,
  show_mark = T,
  color_signif = F,
  color_sigval,
  filter_signif = F,
  cluster_rows = F,
  cluster_cols = F,
  show_colnames = F,
  main = "Stat Results Overview"
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{stat_list}{List of stat names to plot; default NA (for all stat objects in D)}

\item{color_formula}{Expression used to color adjusted p-values; default: sign(fc) * -log10(p.adj)}

\item{met_anno}{rowData() column name to use as column annotation}

\item{signif_expr}{Expression used to indicate significant p-values}

\item{signif_mark}{Expression used to indicate significant p-values}

\item{color_signif}{only color significant values? default: F}

\item{color_sigval}{a vector of length 2 containing positive and negative thresholds of significance}

\item{filter_signif}{show only metabolites significant in one or more results? default: T}

\item{cluster_rows}{cluster by rows? default: F}

\item{cluster_cols}{cluster by cols? default: F}

\item{show_colnames}{should column (metabolite) labels be included in the plot? default: F}
}
\value{
results$output: heatmap of statistical results
}
\description{
Generate a heatamp for a list of statistical results using \code{pheatmap::pheatmap}.
}
\examples{
\dontrun{D <- D \%>\% mt_plots_pheatmap_multstats(met_anno = "SUPER_PATHWAY", cluster_cols = T, cluster_rows = T)}

}
\author{
KC
}
