% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_qc_dilutionplot.R
\name{mt_plots_qc_dilutionplot}
\alias{mt_plots_qc_dilutionplot}
\title{Correlate variable with dilution factors from quotient normalization.}
\usage{
mt_plots_qc_dilutionplot(D, comp, boxpl = T, ggadd = NULL)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{comp}{name of colData colum  (sample annotation) to correlation dilution factors with}

\item{boxpl}{produce boxplot (TRUE, default), or beeswarm plot (FALSE) .... only relevant for factor comparison}

\item{ggadd}{further elements/functions to add (+) to the ggplot object}
}
\value{
$result: plot, comparison with dilution factors
}
\description{
\itemize{
\item for factors: will produce boxplot
\item for quantitative variable: will produce scatter plot
\item requires exactly one previous quotient normalization call in the pipeline
}
}
\examples{
\%>\% mt_plots_qc_dilutionplot(comp="group") \%>\% # compare with 'group' sample annotation

}
\author{
JK
}
