% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_sample_box.R
\name{mt_plots_sample_box}
\alias{mt_plots_sample_box}
\title{Box Plot of samples}
\usage{
mt_plots_sample_box(
  D,
  title = "Sample boxplot",
  legend = T,
  ylabel = "Metabolite concentrations",
  plot_logged = F,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{title}{title of box plot. Default: "Sample boxplot".}

\item{legend}{show legend? Default: T.}

\item{ylabel}{y axis label. Default: "Metabolite concentrations".}

\item{plot_logged}{Show plot logged? Note: plot will still be logged if data have been logged before in pipeline. Default: F.}

\item{ggadd}{Further elements/functions to add (+) to the ggplot object. Default: NULL.}

\item{...}{Additional arguments directly passed to aes() of ggplot.}
}
\value{
$results$output: plot, sample box plot
}
\description{
Create a box plot of the samples. Can be colored by factor.
}
\examples{
\dontrun{## sample boxplot, color by colData 'group' variable, with specific title, on log scale,
... \%>\% mt_plots_sample_box(color=group, title='after quotient normalization', plot_logged=T) \%>\% ...
}

}
\author{
JK
}
