% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_sampleboxplot.R
\name{mt_plots_sampleboxplot}
\alias{mt_plots_sampleboxplot}
\title{Boxplot of samples.}
\usage{
mt_plots_sampleboxplot(
  D,
  plottitle = "Sample boxplot",
  legend = T,
  ylabel = "Metabolite concentrations",
  logged = F,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{plottitle}{title of boxplot, default "Sample boxplot"}

\item{legend}{show legend? default: T}

\item{ylabel}{y axis label, default: "Metabolite concentrations"}

\item{logged}{show plot logged? default: F. note: plot will still be logged if data have been logged before in pipeline.}

\item{ggadd}{further elements/functions to add (+) to the ggplot object}

\item{...}{additional arguments directly passed to aes() of ggplot}
}
\value{
$result: plot, sample boxplot
}
\description{
Can be colored by factor.
}
\examples{
\dontrun{## sample boxplot, color by colData 'group' variable, with specific title, on log scale,
... \%>\% mt_plots_sampleboxplot(color=group, plottitle='after quotient normalization', logged=T) \%>\% ...
}

}
\author{
JK
}
