% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_scatter.R
\name{mt_plots_scatter}
\alias{mt_plots_scatter}
\title{Scatter plots}
\usage{
mt_plots_scatter(
  D,
  x = "x",
  statname,
  correct_confounder,
  metab_filter = p.value < 0.05,
  metab_sort = p.value,
  annotation = "{sprintf('P-value: \%.1e', p.value)}",
  rows,
  cols,
  fitline = T,
  fitline_se = T,
  full.info = F,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{x}{what phenotype (from colData(D)) should be used on x axis, default "x"}

\item{statname}{index of the entry in metadata(D)$results that contains statistic object}

\item{correct_confounder}{confounders to adjust for before plotting, formula notation}

\item{metab_filter}{if given, filter will be applied to data and remaining variables will be labelled in plot, default p.value<0.05}

\item{metab_sort}{if given, arrange will be applied to data variables will be sorted, default p.value}

\item{rows}{number rows of boxplots in $result}

\item{cols}{number columns of boxplots in $result}

\item{fitline}{add fit line? (default: T)}

\item{fitline_se}{add standard error range? (default: T)}

\item{full.info}{add full information of all sample annotations and statistics results to plottable data.frame? makes plotting more flexible but can render SE objects huge. default: F}

\item{ggadd}{further elements/functions to add (+) to the ggplot object}

\item{...}{additional ggplot parameters}
}
\value{
$result: plot, scatter plot
}
\description{
Creates one scatter plot per metabolite based on given sample annotations
}
\examples{
\dontrun{# scatter plots as overview of results with a result already in 'comp'
# color by "age" variable in colData
mt_plots_boxplot(x                  = age,
                 statname           = "comp",
                 correct_confounder = ~BATCH_MOCK,
                 metab_filter       = p.value<0.01,
                 metab_sort         = p.value,
                 annotation         = "{sprintf('P-value: \%.1e', p.value)}\nStatistic: {sprintf('\%.2f', statistic)}",
                 rows               = 2,
                 cols               = 2) \%>\%
}


}
\author{
JK
}
