% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_stats_bar.R
\name{mt_plots_stats_bar}
\alias{mt_plots_stats_bar}
\title{mt_plots_stats_bar}
\usage{
mt_plots_stats_bar(
  D,
  stat_list,
  metab_filter = p.value < 1,
  group_col = "SUB_PATHWAY",
  color_col = NULL,
  y_scale = "fraction",
  sort = FALSE,
  assoc_sign,
  add_empty = FALSE,
  keep_unmapped = FALSE,
  output_file = NULL,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{stat_list}{list of names of the statistics objects to be used for filtering}

\item{metab_filter}{filter will be applied to data and remaining variables will be used to create plot}

\item{group_col}{rowData variable used to aggregate variables.}

\item{color_col}{optional rowData variable used to color barplot. Default NULL.}

\item{y_scale}{plot percentage or frequency of variables. Values c("fraction","count"). Default "fraction".}

\item{sort}{sort pathways in plot according to y_scale. Default FALSE.}

\item{assoc_sign}{optional parameter to discriminate between positive and negative associations. Needs to be the name of a column in the statistical results indicated by stat_list.}

\item{add_empty}{boolean, if TRUE adds also empty pathways to the barplot.}

\item{output_file}{optional Excel filename to save data to}

\item{ggadd}{further elements/functions to add (+) to the ggplot object}

\item{...}{additional expression directly passed to aes() of ggplot, can refer to colData}
}
\value{
$result: plot, barplot
}
\description{
Creates a bar plot
}
\examples{
\dontrun{# Barplot as overview of results with a result already in 'comp'
... \%>\%
mt_plots_stats_bar(stat_list     = "comp",
 metab_filter = p.adj < 0.05,
 group_col    = "SUB_PATHWAY",
 color_col      = "SUPER_PATHWAYdevtools::install(codes.makepath("MT/MetaboTools"))",
 y_scale       = "count",
 sort         = TRUE) \%>\%
 ...}

}
\author{
Elisa Benedetti
}
