% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_stats_compare.R
\name{mt_plots_stats_compare}
\alias{mt_plots_stats_compare}
\title{Comparative plot between two comparisons.}
\usage{
mt_plots_stats_compare(
  D1,
  stat1,
  filter1,
  D2 = D1,
  stat2,
  filter2,
  filter_op = "AND",
  title = "",
  label_col = "name",
  point_size = 1.5,
  return_plot_only = F,
  data_file = NULL,
  use_estimate = F
)
}
\arguments{
\item{D1}{first SE dataset to compare; the one in the pipeline}

\item{stat1}{name of statistical comparison in first dataset}

\item{filter1}{filter term, defining which metabolites to label from first comparison (can use elements of stats table)}

\item{D2}{second SE dataset, if not given, will be the same as the first}

\item{stat2}{name of statistical comparison in second dataset}

\item{filter2}{filter term, defining which metabolites to label from second comparison (can use elements of stats table)}

\item{filter_op}{if AND -> two colors, one for those where both stats match the criterion, and one where they don't
if OR -> three colors, a third one where only one stat matches the criterion}

\item{title}{optional param for plot title}

\item{label_col}{optional argument on which column in the statistical results df to use for labeling points}

\item{point_size}{size of the points on the ggplot}

\item{return_plot_only}{return only the plot object. WARNING: setting this to true makes the function non-MT pipeline compatible.}

\item{data_file}{WHAT DOES THIS DO?}

\item{use_estimate}{use estimate for comparison, instead of statistic; default: F}
}
\value{
$result: plot, p-value histogram
}
\description{
Produces a plot that compares the directed -log10 p-values between two previously executed stats.
}
\examples{
\dontrun{## compare two stats from inside the same pipeline
... \%>\%
mt_plots_stats_compare(stat1='WT',
  filter1= p.adj<0.1,
  stat2='KO',
  filter2= p.adj<0.1,
  filter_op = 'OR'
) \%>\% ...

## compare two stats from different pipelines, as part of the pipeline of the second
# 'comp' is a string that contains the name of a comparison (here both SEs have the same comparison on two datasets)
.. \%>\% mt_plots_stats_compare(
  stat1 = comp, filter1 = p.adj<0.1,
  D2 = firstPipeSE, stat2 = comp, filter2 = p.adj<0.1,
  filter_op = "OR") \%>\% ...

## compare two stats from different pipelines, output as plot object
## not part of the actual MT pipelines, but separate call
# 'comp' is a string that contains the name of a comparison (here both SEs have the same comparison on two datasets)
gg <- mt_plots_stats_compare(
  D1 = D1, stat1 = comp, filter1 = p.adj<0.1,
  D2 = D2, stat2 = comp, filter2 = p.adj<0.1,
  filter_op = "OR", return_plot_only=T)
}

}
\author{
JK
}
