% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_statsbarplot.R
\name{mt_plots_statsbarplot}
\alias{mt_plots_statsbarplot}
\title{mt_plots_statsbarplot}
\usage{
mt_plots_statsbarplot(
  D,
  stat_name,
  metab_filter = p.value < 0.05,
  aggregate = "SUB_PATHWAY",
  colorby = NULL,
  ggadd = NULL,
  yscale = "fraction",
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{stat_name}{optional name of the statistics object to be used for filtering}

\item{metab_filter}{if given, filter will be applied to data and remaining variables will be used to create plot}

\item{aggregate}{rowData variable used to aggregate variables.}

\item{colorby}{optional rowData variable used to color barplot. Default NULL.}

\item{ggadd}{further elements/functions to add (+) to the ggplot object}

\item{yscale}{plot percentage or frequency of variables. Values c("fraction","count"). Default "percentage".}

\item{...}{additional expression directly passed to aes() of ggplot, can refer to colData}
}
\value{
$result: plot, barplot
}
\description{
Creates a bar plot
}
\examples{
\dontrun{# Volcano plot as overview of results with a result already in 'comp'
... \%>\%
mt_plots_statsbarplot(stat_name     = "comp",
 metab_filter = p.adj < 0.05,
 aggregate    = "SUB_PATHWAY",
 colorby      = "SUPER_PATHWAY",
 yscale       = "count") \%>\%
 ...}

}
\author{
Elisa Benedetti
}
