% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_umap.R
\name{mt_plots_UMAP}
\alias{mt_plots_UMAP}
\title{2D UMAP of samples.}
\usage{
mt_plots_UMAP(
  D,
  title = "UMAP",
  scaledata = F,
  labelby = "",
  textrepel = T,
  store.matrices = F,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{title}{Title of the plot, default: "UMAP"}

\item{scaledata}{scale data before plotting? (mean 0, std 1), default: FALSE}

\item{labelby}{field to label. default: none}

\item{textrepel}{try to avoid all text overlaps when labeling? default:T}

\item{store.matrices}{store scores in result structure? default: F}

\item{ggadd}{further elements/functions to add (+) to the ggplot object}

\item{...}{# additional expression directly passed to aes() of ggplot, can refer to colData}
}
\value{
result output: plot(s)

result output2: scores and loadings matrix
}
\description{
Can be colored by any variable in colData.
}
\examples{
\dontrun{## UMAP on scaledata, color and shape by "Group" variable in colData
... $>$ mt_plots_UMAP(scaledata=T, color=Group, shape=Group, title="UMAP - scaled data") \%>\% ...
}

}
\author{
JK
}
