% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_post_addFC.R
\name{mt_post_addFC}
\alias{mt_post_addFC}
\title{Compute fold-change}
\usage{
mt_post_addFC(
  D,
  stat_name,
  correct_confounder,
  combine = function(x) {     mean(x, na.rm = T) }
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{stat_name}{name of the statistical comparison}

\item{correct_confounder}{confounders to adjust for before calculating fold-changes}

\item{combine}{function with respect to which fold changes are computed (e.g. difference of the log means or median)}
}
\value{
$result: statistical object
}
\description{
Add metabolite fold-changes to results table
}
\examples{
\dontrun{# add fold-changes to the result table of the statistical comparison called "comparison1", after correcting for variable "age"
... \%>\%
 mt_post_addFC(stat_name="comparison1", correct_confounder=age) \%>\% ...
 }


}
\author{
JZ
}
