% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_post_multTest.R
\name{mt_post_multTest}
\alias{mt_post_multTest}
\title{Multiple testing correction}
\usage{
mt_post_multTest(D, stat_name, p_col = p.value, method = "bonferroni")
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{stat_name}{name of the statistical comparison to adjust}

\item{p_col}{name of p-value column to adjust}

\item{method}{which method to use for multiple testing (see p.adjust)}
}
\value{
$result: statistical object
}
\description{
Adjust output of statistical tests for multiple testing
}
\examples{
\dontrun{# correct the statistical comparison called "Li's" using Benjamini-Hochberg
... \%>\%
 mt_post_multTest(stat_name="Li's", method="BH") \%>\% ...
 }

}
\author{
JK, JZ
}
