% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_post_pgain.R
\name{mt_post_pgain}
\alias{mt_post_pgain}
\title{Compute p-gain from metabolite ratio test}
\usage{
mt_post_pgain(D, stat_name, p_col = p.value)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{stat_name}{Name of the statistical comparison.}

\item{p_col}{Name of p-value column to compute p-gain from. Default: p.value.}
}
\value{
$results[[stat_name]]$output: pgain column added to statistical table
}
\description{
Add p-gain to result table.
}
\examples{
\dontrun{# add p-gains to the result table of the statistical comparison called "comparison1"
... \%>\%
 mt_post_pgain(stat_name="comparison1") \%>\% ...
 }

}
\author{
JZ
}
