% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_batch_COMBAT.R
\name{mt_pre_batch_COMBAT}
\alias{mt_pre_batch_COMBAT}
\title{COMBAT batch correction.}
\usage{
mt_pre_batch_COMBAT(D, batches)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{batches}{sample annotation (colData) column name that contains batch assignment}
}
\value{
assay: batch-corrected version
}
\description{
Performs batch correction via COMBAT method from \code{sva} package
}
\examples{
\dontrun{... \%>\% mt_pre_batch_COMBAT(batches="BATCH") \%>\% ...
}

}
\author{
JK
}
