% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_batch_median.R
\name{mt_pre_batch_median}
\alias{mt_pre_batch_median}
\title{Median batch correction.}
\usage{
mt_pre_batch_median(D, batches, ref_samples)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{batches}{# sample annotation (colData) column name that contains batch assignment}
}
\value{
assay: batch-corrected version
}
\description{
Same approach that Metabolon uses for runday correction. Divides metabolite values by the median value per batch.
}
\examples{
\dontrun{... \%>\% mt_pre_batch_median(batches="BATCH") \%>\% ...
}

}
\author{
JK
}
