% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_confounding_correction.R
\name{mt_pre_confounding_correction}
\alias{mt_pre_confounding_correction}
\title{Confounding correction}
\usage{
mt_pre_confounding_correction(D, formula, strata = NULL, scale = F)
}
\arguments{
\item{D}{SummarizedExperiment object}

\item{formula}{formula uncluding column names from sample annotation, i.e, ~batch+age}

\item{strata}{strata classes for stratified confounding correction, should be a column from colData(D)}

\item{scale}{should data be scaled after correaction}
}
\value{
corrected data together bit lm.fit pvalue per metabolites to show confounding affect on them
}
\description{
\code{mt_pre_confounding_correction} returns the corrected data
}
\details{
each metabolits is corrected by lm with the confounding variables and residuals are returned
in lieu of uncorrected metabolites expressions.
}
\examples{

 \dontrun{# not run
 #... \%>\% mt_pre_confounding_correction( formula = ~batch + age, strata = "RUN_DAY" )
 }


}
\author{
mubu
}
