% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_confounding_correction_stepwise_aic.R
\name{mt_pre_confounding_correction_stepwise_aic}
\alias{mt_pre_confounding_correction_stepwise_aic}
\title{Confounding correction using stepwise AIC}
\usage{
mt_pre_confounding_correction_stepwise_aic(D, cols_to_cor, n_cores = 1)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{cols_to_cor}{vector of column numbers from colData to correct for}

\item{n_cores}{number of cores to use in parallelization}
}
\value{
SE with corrected data

$output: returns data.frame with the metabolite, its covars and the fit pval and rsq values
}
\description{
Each metabolite is corrected by lm with the confounding variables determined by stepwise aic and residuals are returned
in lieu of uncorrected metabolites expressions.
}
\examples{
 \dontrun{#... \%>\% mt_pre_confounding_correction_stepwise_aic(cols_to_cor = c(1, 4, 5), n_cores = 10)
 }


}
\author{
AS, RB
}
