% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_filter_missingness.R
\name{mt_pre_filter_missingness}
\alias{mt_pre_filter_missingness}
\title{Filter by missingness}
\usage{
mt_pre_filter_missingness(
  D,
  met_max = NA,
  sample_max = NA,
  group_col = NA,
  report_filtered = F,
  report_sample_col = ""
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{met_max}{Maximum fraction of missing metabolites (between 0 and 1.0).}

\item{sample_max}{Maximum fraction of missing samples (between 0 and 1.0).}

\item{group_col}{Name of of a colData sample annotation column; met_max will be applied to each group separately,
metabolite must have at most met_max in any of the groups.}

\item{report_filtered}{Write list of filtered metabolites into the status log text? Default: F.}

\item{report_sample_col}{Required if report_filtered=T for sample filtering. Specifies which column of colData() to
output into the status log.}
}
\value{
assay: Rows or columns will be filtered.

$results$output: Logical vector of kept metabolites/samples.
}
\description{
Filters either samples or metabolites by fraction of missing values. Won't do both in one call.
}
\examples{
\dontrun{# first remove samples with >10\% missingness, then metabolites with >20\% missingness
... \%>\%
  mt_pre_filter_missingness(sample_max=0.1) \%>\%
  mt_pre_filter_missingness(met_max=0.2) \%>\%
...}

}
\author{
JK
}
