% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_impute_knn.R
\name{mt_pre_impute_knn}
\alias{mt_pre_impute_knn}
\title{Impute using kNN method.}
\usage{
mt_pre_impute_knn(D, method = "knn.obs.euc.sel", K = 10, verbose = F)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{K}{Number of nearest neighbors to consider, default is 10 (recommendation: do not touch)}

\item{verbose}{T/F, whether to output intermediate steps, default: F}

\item{methods}{Specific KNN method to use, default is "knn.obs.euc.sel" (recommendation: do not touch)}
}
\value{
assay: imputed data
}
\description{
Default settings are from the winning approach of our paper.
https://link.springer.com/article/10.1007%2Fs11306-018-1420-2
}
\examples{
\dontrun{# in the context of a SE pipeline
... \%>\% mt_pre_impute_knn() \%>\% ...    # standard call
}

}
\author{
JK
}
