% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_impute_knn_multicore.R
\name{mt_pre_impute_knn_multicore}
\alias{mt_pre_impute_knn_multicore}
\title{Impute using kNN method with multi-core functionality.}
\usage{
mt_pre_impute_knn_multicore(D, K = 10, mc_cores = 5, verbose = F)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{K}{Number of nearest neighbors to consider, default is 10 (recommendation: do not touch)}

\item{mc_cores}{Number of cores to use for imputation, dafualt: 5}

\item{verbose}{T/F, whether to output intermediate steps, default: F}
}
\value{
assay: imputed data
}
\description{
Default settings are from the winning approach of Kiki's paper paper.
https://link.springer.com/article/10.1007%2Fs11306-018-1420-2
Specifically "knn.obs.euc.sel"
}
\details{
This script creates a tmp folder where distance matrices and imputed
values are stored
}
\examples{
\dontrun{# in the context of a SE pipeline
... \%>\% mt_pre_impute_knn_multicore() \%>\% ...    # standard call
}

}
\author{
Parviz Gomari
}
