% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_norm_external.R
\name{mt_pre_norm_external}
\alias{mt_pre_norm_external}
\title{Normalize by an external sample annotation}
\usage{
mt_pre_norm_external(D, col_name)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{col_name}{Numeric column in \code{colData} (samples) to normalize by}
}
\value{
assay: normalized version
}
\description{
Normalize by a column in colData, e.g. sample weight, DNA content, protein content (BRADFORD) etc.
}
\examples{
\dontrun{#' # in the context of a SE pipeline
... \%>\% mt_pre_norm_quot(col_name='DNA') \%>\% ...    # normalize by values in column DNA
}

}
\author{
JK
}
