% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_norm_quot.R
\name{mt_pre_norm_quot}
\alias{mt_pre_norm_quot}
\title{Quotient normalization}
\usage{
mt_pre_norm_quot(
  D,
  vars = 1:dim(D)[1],
  na_err = F,
  ref_samples = NULL,
  met_max = 1
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{vars}{index vector of variables of be used, default: all}

\item{na_err}{T/F, throw error for NA's or just ignore?}

\item{ref_samples}{expression filtering reference samples from colData}

\item{met_max}{maximum fraction of missingness to select metabolites to be used in the reference (default: 1, i.e. all metabolites)}
}
\value{
assay: quotient-normalized version

$result: dilution factors
}
\description{
Implementation according to Dieterle et al., 2006\cr
https://www.ncbi.nlm.nih.gov/pubmed/16808434
}
\examples{
\dontrun{#' # in the context of a SE pipeline
... \%>\% mt_pre_norm_quot() \%>\% ...    # standard call
... \%>\% mt_pre_norm_quot(ref_samples = GROUP=="ctrl") \%>\% ...    # use reference samples where 'GROUP' field in colData is 'ctrl'
#' ... \%>\% mt_pre_norm_quot(met_max = 0.2) \%>\% ...    # use only metabolites with <= 20\% missing values to compute the reference used for normalization
}

}
\author{
JK
}
