% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_outlier_detection.R
\name{mt_pre_outlier_detection}
\alias{mt_pre_outlier_detection}
\title{Identifies sample outliers}
\usage{
mt_pre_outlier_detection(D, method = "univariate", reduce_dim = F, ...)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{method}{Can be either "univariate" or "leverage" (for now).}

\item{reduce_dim}{Perform PCA-based dimension reduction before outlier detection? Needed for multivariate methods in low-rank datasets.}

\item{thresh}{Number of standard deviations or m/n units to use as threshold to define the outlier, default value set to 4}

\item{perc}{For the univariate method, percentage of metabolites that need to be outliers in order to consider the whole sample an outlier, default value set to 0.5}

\item{pval}{For the mahalanobis distance method, p-val of chi-squared test to threshold at, default = 0.01}
}
\value{
colData: new columns including a binary vector and a numeric score vector

$results$output: returns the specific parameters used to determine outliers for the method selected
}
\description{
Uses either univariate or multivariate (leverage) approaches, won't do both in one call.
}
\examples{
\dontrun{# first identify samples that have more than 50\% univariate outliers, then identify multivariate outliers with a leverage >4m/n
... \%>\%
  mt_pre_outlier_detection(method="univariate", thresh=4, perc=0.5) \%>\%
  mt_pre_outlier_detection(method="leverage", thresh=4) \%>\%
  mt_pre_outlier_detection(method="mahalanobis", pval=0.01) \%>\%
...}

}
\author{
EB, JK
}
