% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_outliercorrection.R
\name{mt_pre_outliercorrection}
\alias{mt_pre_outliercorrection}
\title{Identifies single outliers in samples.}
\usage{
mt_pre_outliercorrection(
  D,
  threshold = NA,
  sample_num_correction = T,
  alpha = 0.05,
  by_quant = FALSE,
  quant_thresh = 0.025,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{threshold}{Number of standard deviations or m/n units to use as threshold to define the outlier, if not provided, default is correction by sample size. This parameter is only used if sample_num_correction is False}

\item{sample_num_correction}{Whether number of outliers should depend on number of samples or hard sd cutoff. If true, threshold is ignored}

\item{alpha}{The percentage of points we will consider outliers based on the assumption of a normal distribution (alpha/2 on either tail). Only used if sample_num_correction is True}

\item{by_quant}{Optional, if to flag the outliers in the quantile provided in the quantile parameter}

\item{quant_thresh}{If by_quant is true provide a quantile}
}
\value{
SE with NA values where outliers used to be
}
\description{
Uses univariate approach
}
\examples{
\dontrun{... \%>\%
  mt_pre_outliercorrection(threshold=3, sample_num_correction=F) \%>\%
...}

}
\author{
Annalise Schweickart, RB (modified on 2020-05-21)
}
