% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_trans_scale.R
\name{mt_pre_trans_scale}
\alias{mt_pre_trans_scale}
\title{Scale data, mean 0 / sd 1 by default}
\usage{
mt_pre_trans_scale(D, center = T, scale = T, ref_samples)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{center}{T/F, mean-center data? default: T}

\item{scale}{T/F, scale data to sd 1? default: T}

\item{ref_samples}{term which samples to use for center and scale calculcation}
}
\value{
assay: scaled data
}
\description{
This function scales each metabolite by subtracting its mean and dividing by its standard deviation.
Both steps are optional. This function is the direct equivalent of R's scale() function.
}
\examples{
\dontrun{# in the context of a SE pipeline
... \%>\% mt_pre_trans_scale() \%>\% ...    # standard call, center and scale
... \%>\% mt_pre_trans_scale(scale=F) \%>\% ...    # only mean centering
}

}
\author{
JK
}
