% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_heading.R
\name{mt_reporting_heading}
\alias{mt_reporting_heading}
\title{Store heading that will be added to report later.}
\usage{
mt_reporting_heading(D, strtitle, lvl = 1)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{strtitle}{Heading text}

\item{lvl}{Heading level, default: 1 (can be used for nested outline structures)}
}
\value{
$result: stores info about heading
}
\description{
Will be used when calling \code{mt_reporting_generateMD}.
}
\examples{
\dontrun{... \%>\%
# add first and second level of heading
mt_reporting_heading("Preprocessing") \%>\%
mt_reporting_heading("Part 1", lvl=2) \%>\%
...}

}
\author{
JK
}
