% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_hmtl_nonLinear.R
\name{mt_reporting_html_nonLinear}
\alias{mt_reporting_html_nonLinear}
\title{Generates markdown-based HTML output for Non-Linear Pipelines from a list of SummarizedExperiment Objects}
\usage{
mt_reporting_html_nonLinear(
  pipelines,
  outfile,
  title = "RMD output",
  output.calls = F,
  keep.tmp = F
)
}
\arguments{
\item{pipelines}{A list of SummarizedExperiment objects}

\item{outfile}{# output HTML file name}

\item{title}{Title of RMD document}

\item{output.calls}{Output detailed info on function calls? default: F (passed through to mt_reporting_generateMD)}

\item{keep.tmp}{Keep temporary files? (default: No). Can be used to manually edit RMD afterwards.}
}
\description{
Shortcut for mt_reporting_generateMD_nonLinear with subsequent knitting and clean-up.
}
\author{
JK, KC
}
