% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_text.R
\name{mt_reporting_text}
\alias{mt_reporting_text}
\title{Store heading that will be added to report later.}
\usage{
mt_reporting_text(D, text)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{text}{Heading text}
}
\value{
$result: stores info about heading
}
\description{
Will be used when calling \code{mt_reporting_generateMD}.
}
\examples{
\dontrun{... \%>\%
# add first and second level of heading
mt_reporting_text("Preprocessing") \%>\%
mt_reporting_text("Data were first filtered for missing values, then normalized and log-transformed") \%>\%
...}

}
\author{
EB
}
