% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_settings.R
\name{mt_settings}
\alias{mt_settings}
\title{Sets and outputs global pipeline settings}
\usage{
mt_settings(D, settings)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{settings}{list of settings}
}
\value{
Nothing - leaves \code{D} unchanged
}
\description{
Call without parameters to show current settings Call with list of parameters to set settings. Will crash for invalid setting names.
}
\examples{
\dontrun{...  \%>\%
  mt_settings(list(ggplot_fix=F)) \%>\% ... # deactivate ggplot fixing
 }

}
\author{
JK
}
