% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_glmnet.R
\name{mt_stats_glmnet}
\alias{mt_stats_glmnet}
\title{GLMNET with outer and inner cross-validation}
\usage{
mt_stats_glmnet(
  D,
  stat_name,
  outer_k,
  inner_k,
  family = "binomial",
  alpha,
  lambda = NULL,
  rand_seed
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{stat_name}{name under which this comparison will be stored, must be unique to all other statistical results}

\item{outer_k}{number of folds in the outer cross-validation loop}

\item{inner_k}{number of folds in the inner cross-validation loop}

\item{family}{response type; default: binomial}

\item{alpha}{elasticnet mixing parameter}

\item{lambda}{regularization parameter}

\item{rand_seed}{integer to set RNG state so results (including CV partitioning) can be exactly reproduced}
}
\value{
results$ouput: a list containing the foloowing:
\itemize{
 \item{table}{the coefficient matrix returned for glmnet trained on all data}
 \item{lstobj}{the glmnet model trained using on all data}
}

results$output2: a list containing the following:
\itemize{
 \item{idx}{list of indices used to create the train:test partitions for the outer cross-validation loop}
 \item{yhat}{list of all predictions from the outer cross-validation loop}
 \item{labels}{list of test labels from the outer cross-validation loop}
 \item{models}{list of all models from the outer cross-validation loop}
 \item{params}{list containing the original parameters}
}
}
\description{
Runs cv.glment (inner cross-validation) on an outer cross-validation loop. Then runs glmnet on all data.
}
\examples{
# example of how to run function
\dontrun{}

}
\author{
KC
}
