% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_multiv_net_GeneNet.R
\name{mt_stats_multiv_net_GeneNet}
\alias{mt_stats_multiv_net_GeneNet}
\title{Computes partial correlation matrix using the GeneNet estimator.}
\usage{
mt_stats_multiv_net_GeneNet(D, stat_name, sample_filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{stat_name}{name of the correlation matrix}

\item{sample_filter}{term defining which samples to use for GGM calculation (default: all samples)}
}
\value{
original SummarizedExperiment as in input

$output: list of pairwise partial correlation coefficients and pvalues, as well as the corresponding variable names
}
\description{
Implementation according to Schäfer and Strimmer, 2006\cr
https://www.ncbi.nlm.nih.gov/pubmed/16646851
}
\examples{
\dontrun{... \%>\%
  mt_stats_multiv_net_GeneNet(stat_name ="pcor") \%>\%
...}

}
\author{
EB
}
