% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mt_stats_pathway_enrichment_with_differential_analysis.R
\name{mt_stats_pathway_enrichment_with_differential_analysis}
\alias{mt_stats_pathway_enrichment_with_differential_analysis}
\title{Pathway enrichment using different methods}
\usage{
mt_stats_pathway_enrichment_with_differential_analysis(
  D,
  pw_col,
  grp_col,
  ctrl_grp,
  case_grp,
  method = "gage"
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{pw_col}{Column containing pathways IDs.}

\item{grp_col}{Column name of column containing grouping variables.}

\item{ctrl_grp}{Name(s) of control group(s) found in grp_col column.}

\item{case_grp}{Name(s) of case/phenotype group(s) found in grp_col column.}

\item{method}{One of: "gage", "fishers".}
}
\value{
$pathways$enrichment_results: a dataframe containing the pathway enrichment results
}
\description{
This pathway enrichment analysis script calculats differential metabolites either
by a classical t-test or the GAGE package and then performs pathway enrichment
by using Fisher's exact test or GAGE.
}
\details{
Implemented approaches:
1. GAGE: https://bmcbioinformatics.biomedcentral.com/articles/10.1186/1471-2105-10-161
2. Fisher's exact test. Will NOT scale() data before. Data matrix can have NAs.
}
\examples{
\dontrun{\%>\% mt_stats_pathway_enrichment("kegg_db",
                                grp_col = "Group",
                                ctrl_grp = "Vehicle",
                                case_grp = c("treatment2", "treatment1") \%>\%
}


}
\author{
PG
}
