% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_univ_cs.R
\name{mt_stats_univ_cs}
\alias{mt_stats_univ_cs}
\title{Composite score models}
\usage{
mt_stats_univ_cs(D, outcome_col, id_col, stat_name, sample_filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{outcome_col}{name of the column in colData to compare with metabolite}

\item{id_col}{name of the column in colData having patient_ID}

\item{stat_name}{name under which this comparison will be stored, must be unique to all other statistical results}

\item{sample_filter}{term which samples to filter to first... e.g. used if the data contains >2 groups but the user wants to run a two-group comparison}
}
\value{
$result: statistics object
}
\description{
Only support specific composite score association analysis in our COVID project.
}
\examples{
\donttest{
 mt_stats_univ_cs(
   outcome_col     = Group,
   id_col = Subject_ID,
   stat_name         = "composite_score",
   sample_filter = (Group \%in\% c("grp1","grp2")),
 ) \%>\% ...
 }

}
\author{
RB
}
