% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_univ_matrixeqtl.R
\name{mt_stats_univ_matrixeqtl}
\alias{mt_stats_univ_matrixeqtl}
\title{Fast linear models using MatrixEQTL package.}
\usage{
mt_stats_univ_matrixeqtl(D, formula, stat_name, sample_filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{formula}{right-hand side of formula to be put into glm function.}

\item{stat_name}{name under which this comparison will be stored, must be unique to all other statistical results}

\item{sample_filter}{term which samples to filter to first... e.g. used if the data contains >2 groups but the user wants to run a two-group comparison}
}
\value{
$result: statistics object
}
\description{
Substantially fasters than regular lm as implemented in mt_stats_univ_lm.
Only supports standard linear models of the form: outcome ~ metabolite + [covariates]
}
\examples{
\donttest{# run lm with no confounders, "Group" as outcome
# filter to groups "Li_2" and "Li_5"
# name the comparison "Li's"
... \%>\%
 mt_stats_univ_matrixeqtl(
   formula      = ~ Group,
   sample_filter = (Group \%in\% c("Li_2","Li_5")),
   stat_name         = "Li's"
 ) \%>\% ...
 }

}
\author{
JK
}
