% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_univ_wilcox.R
\name{mt_stats_univ_wilcox}
\alias{mt_stats_univ_wilcox}
\title{Wilcox test}
\usage{
mt_stats_univ_wilcox(
  D,
  y,
  stat_name,
  sample_filter,
  paired_flag = F,
  exact_flag = NULL
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{y}{name of the column in colData to compare with metabolite}

\item{stat_name}{name under which this comparison will be stored, must be unique to all other statistical results}

\item{sample_filter}{term which samples to filter to first... e.g. used if the data contains >2 groups but the user wants to run a two-group comparison}

\item{paired_flag}{optional to set the paired flag in wilcox.test function, would be applied for numeric y}

\item{exact_flag}{optional to set the exact flag in wilcox.test function}
}
\value{
$result: statistics object
}
\description{
{ADD DESCRIPTION}
}
\examples{
\donttest{# run lm with no confounders, "Group" as outcome
# filter to groups "Li_2" and "Li_5"
# name the comparison "Li's"
... \%>\%
 mt_stats_univ_wilcox(
   y     = Group,
   sample_filter = (Group \%in\% c("Li_2","Li_5")),
   stat_name         = "Li's"
 ) \%>\% ...
 }

}
\author{
RB
}
