% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stripresults.R
\name{mt_stripresults}
\alias{mt_stripresults}
\title{Remove metadata results from SummarizedExperiment}
\usage{
mt_stripresults(D, strip = "all")
}
\arguments{
\item{D}{SummarizedExperiment input}

\item{strip}{What to remove. Has to be "all" or "plots" (default: "all")}
}
\description{
Removes either the entire result list or all plots from the metadata of a SummarizedExperiment.
Can be used to obtain a more lightweight object for further processing.
}
\details{
strip=="all" will only leave a single result in the object, the one from this function.

strip=="plots" will set all ggplot objects to NULL, still allowing HTML reports to be generated.
}
\examples{
\dontrun{# at the end of a pipeline
... \%>\% mt_stripresults(strip="plots")
...}

}
\author{
JK
}
