% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_write_files_stats.R
\name{mt_write_files_stats}
\alias{mt_write_files_stats}
\title{Export statistical results from pipeline into Excel file.}
\usage{
mt_write_files_stats(
  D,
  file,
  compnames = NULL,
  sort.by.p = F,
  output.dir = T,
  metname = NULL
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{file}{output filename to write to}

\item{compnames}{Names of one or more statistical comparison to be written out. If NULL, will export all.}

\item{sort.by.p}{Automatically sort by p-values? (default: False)}

\item{output.dir}{If the statistical test was a 2-group test, output the direction of the effect (which of the two groups the effect was "high" in). Default: Yes}

\item{metname}{#optional parameter with name of rowdata column to include in the output}
}
\value{
Does not change the SummarizedExperiment.
}
\description{
Writes out the statistics data frame with p-values, adjusted p-values, fold changes, test statistics etc. into an Excel sheet.
}
\examples{
\dontrun{# Write out all results
... \%>\% mt_write_files_stats(file="results.xlsx") \%>\%
# Write out specific result]
... \%>\% mt_write_files_stats(file="results.xlsx", compnames="comp1") \%>\%}

}
\author{
JK, RB (modified on 2020-07-22)
}
