% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_write_files_xls.R
\name{mt_write_files_xls}
\alias{mt_write_files_xls}
\title{Outputs assay, colData and rowData into an Excel file.}
\usage{
mt_write_files_xls(D, file)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{file}{output filename to write to}
}
\value{
Does not change the SummarizedExperiment.
}
\description{
Exports the current SummarizedExperiment (not the metadata) to an Excel file.
}
\examples{
\dontrun{\%>\% mt_write_files_xls(file = "out.xlsx") \%>\%}
}
\author{
JK, BGP
}
