% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_write_stats.R
\name{mt_write_stats}
\alias{mt_write_stats}
\title{Export statistical results from pipeline into an Excel file}
\usage{
mt_write_stats(D, file, stat_names = NULL, sort_by_p = F, met_name = NULL)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input.}

\item{file}{Output filename to write to.}

\item{stat_names}{Names of one or more statistical comparison to be written out. If NULL, will export all.}

\item{sort_by_p}{Automatically sort by p-values? Default: False.}

\item{met_name}{OPTIONAL. Name of rowdata column to include in the output.}
}
\value{
Does not change the \code{SummarizedExperiment} object.
}
\description{
Writes out the statistics data frame with p-values, adjusted p-values, fold changes, test statistics etc. into an Excel sheet.
}
\examples{
\dontrun{# Write out all results
... \%>\% mt_write_stats(file="results.xlsx") \%>\%
# Write out specific result]
... \%>\% mt_write_stats(file="results.xlsx", stat_names="comp1") \%>\%}

}
\author{
JK, RB
}
