% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_res_get_path}
\alias{mtm_res_get_path}
\title{Extract results from a "path" of entries in metadata}
\usage{
mtm_res_get_path(D, path)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{path}{path, list of strings}
}
\value{
SummarizedExperiment with corrected data
}
\description{
Extracts metadata()$results entries in a given namespace of arbitrary depth. See examples below.
}
\examples{
\dontrun{# get all result entries for plots
D \%>\% mtm_res_get_path("plots")
# get all result entries for stats
D \%>\% mtm_res_get_path("stats")
# check if quotient normalization has been performed
q <- D\%>\% mtm_res_get_path(c("pre","norm","quot"))
if (length(q)==0) stop("No quotient normalization performed.")
}

}
\author{
JK
}
