% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_res_get_plots_entries}
\alias{mtm_res_get_plots_entries}
\title{Extract all "plots" $results entries from pipeline}
\usage{
mtm_res_get_plots_entries(D)
}
\arguments{
\item{D}{SummarizedExperiment}
}
\description{
Extract all "plots" $results entries from pipeline
}
