% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtm_access.R
\name{mtm_res_get_stats_entries}
\alias{mtm_res_get_stats_entries}
\title{Extract all stats entries.}
\usage{
mtm_res_get_stats_entries(D)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}
}
\value{
list if stats results
}
\description{
Returns all stats entries from the $results list.
}
\author{
JK
}
