% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtw_apply_analysis.R
\name{mtw_apply_analysis}
\alias{mtw_apply_analysis}
\title{mtw_apply_analysis}
\usage{
mtw_apply_analysis(D, outcomes)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{outcomes}{A data frame, should contain two columns: outcomes$outcome: name of phenotype columns in colData; outcomes$type: data type of pheotype columns, must be one of numeric / binary / ordinal}
}
\value{
D with multiple analysis results and plots
}
\description{
Running standard analyses for multiple phenotypes, analysis steps including lm/cor analysis, 
multi-test with histgram, QQ-plot and volcano plot showing p-values, and box/scatter plot for the 
visualization of significant metabolites.
}
\examples{
\dontrun{... \%>\% mtw_apply_analysis(outcomes) \%>\% ...}

}
\author{
Zeyu Wang
}
