% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtw_missingness.R
\name{mtw_missingness}
\alias{mtw_missingness}
\title{Wrapper function to filter missingness}
\usage{
mtw_missingness(D, plot_options = list(), filter_options = list())
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{plot_options}{A list of parameters for \code{mt_plots_qc_missingness}}

\item{filter_options}{A list of oparameters for \code{mt_pre_filtermiss}, default thresholds for metabolites and samples are 0.2 and 0.1}
}
\value{
D with filtered data
 column "Met_Missing" will be added to rowData with missing information on metabolites
 column "Sample_Missing" will be added to colData with missing information on samples
}
\description{
Filtering missingness of both metabolites and samples, and generating 3 missingness plots: before filtering, after filter metabolites, and after filter samples
}
\examples{
\dontrun{... \%>\% mtw_missingness() \%>\% ...}
\dontrun{... \%>\% mtw_missingness(filter_options = list(met_max = 0.1, sample_max = 0.05)) \%>\% ...}

}
\author{
Zeyu Wang
}
