% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtw_preprocess.R
\name{mtw_preprocess}
\alias{mtw_preprocess}
\title{mtw_preprocess}
\usage{
mtw_preprocess(
  D,
  batch_col,
  batch_fun = mt_pre_batch_median,
  batch_ref_samples = NULL,
  boxplot_options = list(),
  quot_options = list(),
  comp_fac = c(),
  dilution_options = list(),
  log_base = 2,
  do_impute = T,
  knn_options = list()
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{batch_col}{Batch column if needed}

\item{batch_fun}{Batch function to use. Default is \code{mt_pre_batch_median}}

\item{boxplot_options}{A list of parameters for \code{mt_plots_sampleboxplot}}

\item{quot_options}{A list of parameters for \code{mt_pre_norm_quot}}

\item{comp_fac}{List of names in colData column to correlation dilution factors with}

\item{dilution_options}{A list of parameters for \code{mt_plots_qc_dilutionplot}}

\item{log_base}{Base for log function, default is 2}

\item{do_impute}{Logical. Whether to run imputation}

\item{knn_options}{A list of parameters for \code{mt_pre_impute_knn}}
}
\value{
D that has been preprocessed
}
\description{
Preprocess on data, including batch correction, quotient normalization, log transformation and kNN imputations
}
\examples{
\dontrun{... \%>\% mtw_preprocess() \%>\% ...}
\dontrun{... \%>\% mtw_preprocess(batch_col = "BATCH_MOCK",comp_fac = c('num1','Group'), batch_col = "BATCH_MOCK",comp_fac = c("num1","Group"),quot_options = list(ref_samples = quote(Group == 'Vehicle'))) \%>\% ...}

}
\author{
Zeyu Wang
}
