% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_metabolites_HMDB2KEGG.R
\name{mt_anno_metabolites_HMDB2KEGG}
\alias{mt_anno_metabolites_HMDB2KEGG}
\title{Create KEGG identifiers from HMDB identifiers.}
\usage{
mt_anno_metabolites_HMDB2KEGG(D, in_col, out_col = "KEGG")
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{in_col}{string name of the rowData column containing the HMDB identifiers}

\item{out_col}{string name of the new rowData column to be created with the KEGG identifiers; default="KEGG"}
}
\value{
rowData: new identifier annotation column added
}
\description{
Creates annotations and merges them into current SummarizedExperiment.
Performs "left-joins", i.e. leaves the original SE unchanged and just adds information where it can be mapped.
}
\examples{
# Load data, sheet with sample annotations, sheet with metabolite annotations and add KEGG identifiers
\dontrun{D <-
  # load raw data
  mt_files_data_xls(file=file, sheet="data", samplesInRows=T, ID="SAMPLE_NAME") \%>\%
  # sample annotations from metabolomics run
  mt_files_anno_xls(file=file, sheet="sampleinfo", annosfor="samples", IDanno = "SAMPLE_NAME") \%>\%
  # metabolite annotations
  mt_files_anno_xls(file=file, sheet="metinfo", annosfor="metabolites", IDanno="BIOCHEMICAL", IDdata = "name") \%>\%
  # add KEGG identifiers
  mt_anno_metabolites_HMDB2KEGG(in_col="HMDb_id", out_col="KEGG") \%>\%
  ...}

}
\author{
EB
}
