% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_anno_pathways_Uniprot.R
\name{mt_anno_pathways_Uniprot}
\alias{mt_anno_pathways_Uniprot}
\title{Add pathway information.}
\usage{
mt_anno_pathways_Uniprot(D, in_col, out_col, export_raw_db)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{in_col}{column to use for pathway fetching. The selected column must contain protein uniprot identifier}

\item{out_col}{new column name for rowData to output pathway information to}

\item{export_raw_db}{OPTIONAL. Export the pathway database to a directory. Must be a string containing the path name with a .xlsx extension.}
}
\value{
rowData: new pathway annotation for proteins

$result$pathways: a dataframe of pathway information
}
\description{
Modified from mt_anno_pathways_HMDB.
 Adds custom pathways to the already existing SummarizedExperiment
 data taken from KEGGREST.
}
\examples{
# annotate proteins using kegg
\dontrun{... \%>\%
  mt_anno_pathways_Uniprot(in_col = "Uniprot_ID", out_col = "kegg_db") \%>\%
...}

}
\author{
Richa Batra
}
