% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_load_nightingale.R
\name{mt_files_load_nightingale}
\alias{mt_files_load_nightingale}
\title{Load Nightingale-format data.}
\usage{
mt_files_load_nightingale(
  excel_file,
  raw_sheet = "Results",
  met_sheet = "Biomarker annotations",
  met_qc = "Tags per biomarker",
  sample_qc = "Quality control tags and notes",
  sample_id = "Sample id"
)
}
\arguments{
\item{excel_file}{input Excel file}

\item{raw_sheet}{name of sheet with raw data}

\item{met_sheet}{name of sheet with biomarker information}

\item{met_qc}{name of sheet with biomarker qc tags}

\item{sample_qc}{name of sheet with sample qc tags}

\item{sample_id}{optional, name of the column with sample id}
}
\value{
Produces an initial SummarizedExperiment, with assay, colData, rowData, and metadata with first entry
}
\description{
Loads data from a Nightingale format Excel file.
}
\examples{
\dontrun{D <-
  # load data
  mt_files_load_nightingale(excel_file = codes.makepath("Mt/sampledata_night.xlsx"), 
  raw_sheet = "Results", 
  met_sheet = "Biomarker annotations",
  met_qc = "Tags per biomarker",
  sample_qc = "Quality control tags and notes",
  sample_id = 'Sample id'# optional name of the column with sample id, default is 'Sample id'
  ) \%>\%
  ...}

}
\author{
RB
}
