% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_files_load_metabolon_new_format.R
\name{mt_load_metabolon_v2}
\alias{mt_load_metabolon_v2}
\title{Load NEW Metabolon-format data}
\usage{
mt_load_metabolon_v2(file, raw_sheet, met_sheet, sample_sheet)
}
\arguments{
\item{file}{Name of input Excel file.}

\item{raw_sheet}{name of sheet with raw data}

\item{met_sheet}{name of sheet with metabolite info}

\item{sample_sheet}{name of sheet with sample info}
}
\value{
Produces an initial SummarizedExperiment, with assay, colData, rowData, and metadata with first entry.
}
\description{
Loads data from a NEW Metabolon-format Excel file. Needs to be in the original "Client Data Table" new format that they deliver.
}
\examples{
\dontrun{D <-
  # load data
  mt_files_load_metabolon_new_format(codes.makepath("Mt/sampledata.xlsx"),
  raw_sheet="Raw Data", met_sheet="Chemical Annotation", sample_sheet="Sample Meta Data") \%>\%
  ...}

}
\author{
RB
}
