% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_modify_icc.R
\name{mt_modify_icc}
\alias{mt_modify_icc}
\title{mt_modify_icc}
\usage{
mt_modify_icc(D, qc_samples, col_lab, grp_col)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{qc_samples}{Logical expression. Can use fields from \code{colData()}}

\item{col_lab}{Label for the new icc values to be stored in in rowData}

\item{grp_col}{it is the name of the column in colData containing groups of qc_samples}
}
\value{
rowData(D)[[col_lab]] contains column with ICC score for each metabolite
}
\description{
Calculate ICC - intraclass correlation
}
\examples{
\dontrun{... \%>\% mt_modify_icc(qc_samples=="PQC", col_lab = "PQC_cv", grp_col='dup_grp') \%>\% ...}

}
\author{
RB
}
