% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_qc_missingness.R
\name{mt_plots_qc_missingness}
\alias{mt_plots_qc_missingness}
\title{Visualize missing value structure of dataset.}
\usage{
mt_plots_qc_missingness(
  D,
  metMax = NA,
  sampMax = NA,
  plot.mets = T,
  plot.samples = F,
  sample.labels = NA,
  plot.data = T
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{metMax}{Which missingness to mark on the y axis (default: NA = no line)}
}
\value{
$result: plot, two plots
}
\description{
Creates two plots:
1. missing value matrix view
2. sorted missing value % plot
}
\examples{
\%>\%  mt_plots_qc_missingness() \%>\% # without horizontal line
\%>\%  mt_plots_qc_missingness(metMax=0.5) \%>\% # with horizontal line at 50\%

}
\author{
JK
}
