% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_statsbarplot.R
\name{mt_plots_statsbarplot}
\alias{mt_plots_statsbarplot}
\title{mt_plots_statsbarplot}
\usage{
mt_plots_statsbarplot(
  D,
  stat_name,
  metab_filter = p.value < 1,
  aggregate = "SUB_PATHWAY",
  colorby = NULL,
  ggadd = NULL,
  yscale = "fraction",
  sort = FALSE,
  assoc_sign,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{stat_name}{list of names of the statistics objects to be used for filtering}

\item{metab_filter}{filter will be applied to data and remaining variables will be used to create plot}

\item{aggregate}{rowData variable used to aggregate variables.}

\item{colorby}{optional rowData variable used to color barplot. Default NULL.}

\item{ggadd}{further elements/functions to add (+) to the ggplot object}

\item{yscale}{plot percentage or frequency of variables. Values c("fraction","count"). Default "fraction".}

\item{sort}{sort pathways in plot according to yscale. Default FALSE.}

\item{assoc_sign}{optional parameter to discriminate between positive and negative associations. Needs to be the name of a column in the statistical results indicated by stat_name.}

\item{...}{additional expression directly passed to aes() of ggplot, can refer to colData}
}
\value{
$result: plot, barplot
}
\description{
Creates a bar plot
}
\examples{
\dontrun{# Barplot as overview of results with a result already in 'comp'
... \%>\%
mt_plots_statsbarplot(stat_name     = "comp",
 metab_filter = p.adj < 0.05,
 aggregate    = "SUB_PATHWAY",
 colorby      = "SUPER_PATHWAYdevtools::install(codes.makepath("MT/MetaboTools"))",
 yscale       = "count",
 sort         = TRUE) \%>\%
 ...}

}
\author{
Elisa Benedetti
}
