% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_plots_volcano.R
\name{mt_plots_volcano}
\alias{mt_plots_volcano}
\title{mt_plots_volcano}
\usage{
mt_plots_volcano(
  D,
  x = fc,
  statname,
  metab_filter = p.value < 0.05,
  xlabel = gsub("~", "", as.character(x)),
  vline = NA,
  hline,
  ggadd = NULL,
  ...
)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{x}{what value shall be plotted on x (default: fc)}

\item{statname}{name of the statistics object to plot}

\item{metab_filter}{if given, filter will be applied to data and remaining varaibles will be labelled in plot}

\item{vline}{where to draw vertical line (for fold-change), has to be single value. default: none}

\item{hline}{where to draw horizontal line (for p-values), has to be an expression such as 'p.adj < 0.1'. default: none}

\item{ggadd}{further elements/functions to add (+) to the ggplot object}

\item{...}{additional expression directly passed to aes() of ggplot, can refer to colData}
}
\value{
$result: plot, volcano
}
\description{
Creates a volcano plot
}
\examples{
\dontrun{# Volcano plot as overview of results with a result already in 'comp'
... \%>\%
mt_plots_volcano(statname     = "comp",
 metab_filter = p.adj < 0.1,
 colour       = p.value < 0.05) \%>\%
 ...}

}
\author{
Jonas Zierer, Jan Krumsiek
}
