% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_pre_batch_COMBAT.R
\name{mt_pre_batch_ComBat}
\alias{mt_pre_batch_ComBat}
\title{ComBat batch correction}
\usage{
mt_pre_batch_ComBat(D, batches)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{batches}{sample annotation (colData) column name that contains batch assignment}
}
\value{
assay: batch-corrected version
}
\description{
Performs batch correction via ComBat method from \code{sva} package.
}
\examples{
\dontrun{... \%>\% mt_pre_batch_ComBat(batches="BATCH") \%>\% ...
}

}
\author{
JK
}
