% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_html.R
\name{mt_reporting_html}
\alias{mt_reporting_html}
\title{Generates markdown-based HTML output from SummarizedExperiment}
\usage{
mt_reporting_html(
  D,
  outfile,
  title = "RMD output",
  output.calls = F,
  start.after = NA,
  use.plotly = F
)
}
\arguments{
\item{D}{Summarized experiment input}

\item{outfile}{# output HTML file name}

\item{title}{Title of RMD document}

\item{output.calls}{Output detailed info on function calls? default: F (passed through to mt_reporting_generateMD)}

\item{start.after}{UUID of pipeline step AFTER which to start (default: none, i.e. output entire pipeline) (passed through to mt_reporting_generateMD)}

\item{use.plotly}{Output interactive plotly plots? (experimental)}
}
\description{
Shortcut for mt_reporting_generateMD with subsequent knitting and clean-up.
}
\author{
JK
}
