% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_hmtl_nonLinear.R
\name{mt_reporting_html_nonlinear}
\alias{mt_reporting_html_nonlinear}
\title{Generates markdown-based HTML output for non-linear pipelines from a list of SummarizedExperiment objects}
\usage{
mt_reporting_html_nonlinear(
  D_list,
  outfile,
  title = "Non-Linear RMD output",
  output_calls = F,
  keep_tmp = F
)
}
\arguments{
\item{D_list}{A list of \code{SummarizedExperiment} objects.}

\item{outfile}{Output HTML filename.}

\item{title}{Title of RMD document. Default: 'Non-Linear RMD output'.}

\item{output_calls}{Output detailed info on function calls? Default: F.}

\item{keep_tmp}{Keep temporary files? Can be used to manually edit RMD afterwards. Default: F.}
}
\value{
Nothing. This does not pass through \code{SummarizedExperiment} objects.
}
\description{
Shortcut for mt_reporting_generateMD_nonLinear with subsequent knitting and clean-up.
}
\author{
JK, KC
}
