% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_reporting_text.R
\name{mt_reporting_text}
\alias{mt_reporting_text}
\title{Store text that will be added to html report later.}
\usage{
mt_reporting_text(D, text)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{text}{Heading text}
}
\value{
$result: stores plain text for use with html
}
\description{
Plain text to add to the html document. Will be used when calling \code{mt_reporting_html} or \code{mt_reporting_html_nonLinear}.
}
\examples{
\dontrun{... \%>\%
# add first and second level of heading
mt_reporting_heading("Preprocessing") \%>\%
mt_reporting_text("Data were first filtered for missing values, then normalized and log-transformed") \%>\%
...}

}
\author{
EB
}
