% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_stats_univ_cs.R
\name{mt_stats_univ_cs}
\alias{mt_stats_univ_cs}
\title{Composite score models
Only support specific composite score association analysis in our COVID project}
\usage{
mt_stats_univ_cs(D, y, patient_id, stat_name, sample_filter)
}
\arguments{
\item{D}{\code{SummarizedExperiment} input}

\item{y}{name of the column in colData to compare with metabolite}

\item{patient_id}{name of the column in colData having patient_ID}

\item{stat_name}{name under which this comparison will be stored, must be unique to all other statistical results}

\item{sample_filter}{term which samples to filter to first... e.g. used if the data contains >2 groups but the user wants to run a two-group comparison}
}
\value{
$result: statistics object
}
\description{
Composite score models
Only support specific composite score association analysis in our COVID project
}
\examples{
\donttest{
 mt_stats_univ_cs(
   y     = Group,
   patient_id = Subject_ID,
   stat_name         = "composite_score",
   sample_filter = (Group \%in\% c("grp1","grp2")),
 ) \%>\% ...
 }

}
\author{
RB
}
